% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{allprobs}
\alias{allprobs}
\title{Density values of observations in states}
\usage{
allprobs(observations, sdd, state, ...)
}
\arguments{
\item{observations}{A \code{numeric} \code{vector} of observations.}

\item{sdd}{A single specification for the \code{"sdds"} control,
see \code{\link{set_controls}}.}

\item{state}{A single specification for the \code{"states"} control,
see \code{\link{set_controls}}.}

\item{...}{Vectors with parameters for each state, named \code{mu}, \code{sigma}, and
\code{df}, for the mean, standard deviation, and degrees of freedom,
respectively (if required).}
}
\value{
A \code{matrix} with \code{state} rows and \code{length(observations)}
columns. The element in row \eqn{i} and column \eqn{j} is the density value
for observation \eqn{j} in state \eqn{i}.
}
\description{
This helper function computes all density values for all observations in all
possible states.
}
\keyword{internal}
