% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_viterbi.R
\name{apply_viterbi}
\alias{apply_viterbi}
\title{Viterbi algorithm}
\usage{
apply_viterbi(data, fit, controls)
}
\arguments{
\item{data}{A list of processed data information.}

\item{fit}{A list of fitted model information.}

\item{controls}{A list of controls.}
}
\value{
A vector (in case of a HMM) or a matrix (in case of a hierarchical HMM) of decoded states.
}
\description{
Applies the Viterbi algorithm (\url{https://en.wikipedia.org/wiki/Viterbi_algorithm}) for state decoding.
}
