% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{build_draw_function}
\alias{build_draw_function}
\title{Build \code{function} to draw from state-dependent distribution}
\usage{
build_draw_function(class, fixed_parameters)
}
\arguments{
\item{class}{One of \code{"norm"}, \code{"t"}, \code{"gamma"}, \code{"lnorm"},
\code{"poisson"}.}

\item{fixed_parameters}{A named \code{list} of fixed parameters for the state-dependent distribution.
In most cases, it is the output of \code{\link{clean_fixed_parameters}}.}
}
\value{
A \code{function} that returns random draws. It has the following arguments:
\itemize{
\item \code{n}, the number of draws (by default \code{n = 1}),
\item \code{mean}, \code{sd}, and \code{df} for the mean, standard
deviation, and degrees of freedom, respectively (if required),
in each state,
\item \code{state}, the state number.
}
The input \code{fixed_parameters} is used to define default arguments.
}
\description{
This helper function builds a \code{function} that can draw from a
state-dependent distribution.
}
\examples{
\dontrun{
draw_t_sdd <- build_draw_function("t", list("sd" = 1:2, "df" = 1:2))
draw_t_sdd(mean = c(-10, 10), state = 1)
draw_t_sdd(mean = c(-10, 10), state = 2)
}

}
\keyword{internal}
