% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{build_sample_function}
\alias{build_sample_function}
\title{Build sampling function for state-dependent distribution}
\usage{
build_sample_function(distr_class)
}
\arguments{
\item{distr_class}{A \code{character} which defines the class of the state-dependent
distribution class and can be one of
\itemize{
\item \code{"norm"} (normal distribution),
\item \code{"t"} (t-distribution),
\item \code{"gamma"} (gamma distribution),
\item \code{"lnorm"} (log-normal distribution),
\item \code{"poisson"} (poisson distribution),
}}
}
\value{
A \code{function} which returns random samples from the state-dependent
distribution and has the arguments
\itemize{
\item \code{n}, the number of samples (by default \code{n = 1}),
\item \code{mu}, \code{sigma}, and \code{df}, vectors for the mean,
standard deviation, and degrees of freedom, respectively
(if required), in each state,
\item \code{state}, the state index.
}
}
\description{
This helper function builds a \code{function} that can sample from a
state-dependent distribution.
}
\examples{
\dontrun{
sample_poisson_sdd <- build_sample_function("poisson")
mu <- c(1, 10)
sample_poisson_sdd(mu = mu, state = 1)
sample_poisson_sdd(mu = mu, state = 2)
}

}
\keyword{internal}
