% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{build_sdd_function}
\alias{build_sdd_function}
\title{Build sampling, density, or cumulative distribution function for
state-dependent distributions}
\usage{
build_sdd_function(distr_class, function_type)
}
\arguments{
\item{distr_class}{A \code{character} which defines the class of the state-dependent
distribution, see the output of \code{\link{decode_sdd}}.}

\item{function_type}{A \code{character} which defines the desired function type and can be one of
\itemize{
\item \code{"sample"} (sampling function),
\item \code{"density"} (density function),
\item \code{"distribution"} (distribution function).
}}
}
\value{
A \code{function} which either returns
\itemize{
\item random samples if \code{function_type = "sample"},
\item the density if \code{function_type = "density"},
\item the cumulative distribution if \code{function_type = "distribution"},
}
from the state-dependent distribution specified as \code{distr_class}.

The returned \code{function} has the arguments
\itemize{
\item either \itemize{
\item \code{n}, the number of samples (by default \code{n = 1}),
\item \code{x}, the point where to compute the density,
\item \code{q}, the point where to compute the cumulative distribution,
}
\item \code{state}, the state index,
\item \code{...}, vectors with parameters for each state, named
\code{mu}, \code{sigma}, and \code{df}, for the mean,
standard deviation, and degrees of freedom, respectively
(if required).
}
}
\description{
This helper function builds a \code{function} for sampling, density
computation, or cumulative distribution for a state-dependent distribution.
}
\keyword{internal}
