% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{check_all_list_elements_named}
\alias{check_all_list_elements_named}
\title{Check if all elements of a \code{list} are named}
\usage{
check_all_list_elements_named(x)
}
\arguments{
\item{x}{A \code{list}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
This helper function checks whether all elements of a \code{list} are named.
}
\examples{
\dontrun{
check_all_list_elements_named(list())
check_all_list_elements_named(list(1))
check_all_list_elements_named(list("a" = 1))
check_all_list_elements_named(list("a" = 1, 2))
check_all_list_elements_named(list("a" = 1, "b" = 2))
}

}
\keyword{internal}
