% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_controls.R
\name{check_controls}
\alias{check_controls}
\title{Check controls}
\usage{
check_controls(controls)
}
\arguments{
\item{controls}{A list of controls.}
}
\value{
Checked version of \code{controls}.
}
\description{
This function checks the specification of \code{controls}.
}
\details{
See the vignettes on how to specify \code{controls}.
}
