% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_date}
\alias{check_date}
\title{Check date format}
\usage{
check_date(date)
}
\arguments{
\item{date}{[\code{character(1)}]\cr
The date in format \code{"YYYY-MM-DD"}.}
}
\value{
\code{as.Date(date)} if \code{date} has the format \code{"YYYY-MM-DD"}.
Otherwise, the function throws an error.
}
\description{
This function checks if the input \code{date} has the format
\code{"YYYY-MM-DD"}.
}
\keyword{internal}
