% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_decoding.R
\name{check_decoding}
\alias{check_decoding}
\title{Decoding check}
\usage{
check_decoding(decoding, data, controls)
}
\arguments{
\item{decoding}{A vector (in case of a hmm) or a matrix (in case of a hierarchical HMM) of decoded states.}

\item{data}{A list of processed data information.}

\item{controls}{A list of controls.}
}
\value{
No return value. Creates output file "states.txt".
}
\description{
Summarizes and saves decoded states.
}
