% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimation.R
\name{check_estimation}
\alias{check_estimation}
\title{Estimation check}
\usage{
check_estimation(mods, lls, data, hessian, controls)
}
\arguments{
\item{mods}{A list of fitted models in the different estimation runs.}

\item{lls}{A vector of log-likelihood values of accepted \code{mods}.}

\item{data}{A list of processed data information.}

\item{hessian}{Hessian matrix of the estimated model.}

\item{controls}{A list of controls.}
}
\value{
A list of fitted model information.
}
\description{
Summarizes and saves estimates.
}
