% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{clean_fixed_parameters}
\alias{clean_fixed_parameters}
\title{Clean fixed parameters for state-dependent distribution}
\usage{
clean_fixed_parameters(class, nstates, ...)
}
\arguments{
\item{class}{One of \code{"norm"}, \code{"t"}, \code{"gamma"}, \code{"lnorm"},
\code{"poisson"}.}

\item{nstates}{An \code{integer}, the number of states.}

\item{...}{TODO}
}
\value{
A \code{list} ... TODO
}
\description{
This helper function checks the fixed parameters for a state-dependent
distribution.
}
