% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Comparison of multiple \code{fHMM_model}.}
\usage{
compare_models(...)
}
\arguments{
\item{...}{A list of one or more objects of class \code{fHMM_model}.}
}
\value{
A data frame.
}
\description{
This function compares multiple \code{fHMM_model}.
}
