% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fs.R
\name{compute_fs}
\alias{compute_fs}
\title{Fine-scale chunk lengths}
\usage{
compute_fs(fs_time_horizon, T = NA, fs_dates = NA)
}
\arguments{
\item{fs_time_horizon}{Either a numeric or one of \code{"w"}, \code{"m"}, \code{"q"}, \code{"y"}, setting the fine-scale dimension.}

\item{T}{A numeric, the dimension of the coarse-scale process, default \code{NA}.}

\item{fs_dates}{A vector of dates of empirical fine-scale observations, default \code{NA}.}
}
\value{
A vector of fine-scale chunk sizes.
}
\description{
Computes (flexible) fine-scale chunk lengths.
}
