% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{dax}
\alias{dax}
\title{DAX data}
\format{
A \code{data.frame} with the columns:
TODO
}
\usage{
dax
}
\description{
DAX data from ... to ... scraped from ...
TODO
}
\details{
The data was obtained via ...
TODO
}
\keyword{data}
