% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{decode_sdd}
\alias{decode_sdd}
\title{Decode specification for state-dependent distribution}
\usage{
decode_sdd(sdd, states)
}
\arguments{
\item{sdd}{A single specification for the \code{"sdds"} control,
see \code{\link{set_controls}}.}

\item{states}{A single specification for the \code{"states"} control,
see \code{\link{set_controls}}.}
}
\value{
A \code{list} with the elements
\itemize{
\item \code{"distr_class"}, a \code{character} which defines the
distribution class and can be one of
\itemize{
\item \code{"normal"} (normal distribution),
\item \code{"lognormal"} (log-normal distribution),
\item \code{"t"} (t-distribution),
\item \code{"gamma"} (gamma distribution),
\item \code{"poisson"} (poisson distribution),
}
\item \code{"label"}, a \code{character} label for the specified
distribution,
\item \code{"fixed_pars"}, a \code{list} of fixed parameter values,
\item \code{"sample"}, a \code{function} to sample from the state-dependent
distribution,
\item \code{"density"}, a \code{function} to compute the density of the
state-dependent distribution,
\item \code{"distribution"}, a \code{function} to compute the cumulative
density of the state-dependent distribution.
}
}
\description{
This helper function decodes the specification for a state-dependent
distribution.
}
\examples{
\dontrun{
decode_sdd("t(mu = 1)", states = 3)
}

}
\keyword{internal}
