% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCon2dfUncon.R
\name{dfCon2dfUncon}
\alias{dfCon2dfUncon}
\title{Unconstrains degrees of freedom.}
\usage{
dfCon2dfUncon(dfCon)
}
\arguments{
\item{dfCon}{Numeric, constrained degrees of freedom.}
}
\value{
Numeric, unconstrained degrees of freedom.
}
\description{
Unconstrains degrees of freedom.
}
