% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{dfUncon2dfCon}
\alias{dfUncon2dfCon}
\title{This function constrains the un-constrained degrees of freedom \code{dfUncon}.}
\usage{
dfUncon2dfCon(dfUncon)
}
\arguments{
\item{dfUncon}{A vector of un-constrained degrees of freedom.}
}
\value{
A vector of constrained degrees of freedom.
}
\description{
This function constrains the un-constrained degrees of freedom \code{dfUncon}.
}
\keyword{internal}
