% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_controls.R
\name{.simulated_data}
\alias{.simulated_data}
\title{Helper function that checks if data was simulated}
\usage{
.simulated_data(x)
}
\arguments{
\item{x}{Either a \code{list} or an object of class \code{fHMM_controls}.}
}
\description{
Helper function that checks if data was simulated
}
\keyword{internal}
