% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{exception}
\alias{exception}
\title{Exception messages}
\usage{
exception(code)
}
\arguments{
\item{code}{An exception code}
}
\value{
A list containing the following elements:
\item{code}{the exception code}
\item{response}{the message}
\item{debugging}{suggestions for debugging}
}
\description{
Exception messages
}
\examples{
exception("S.1")
}
