% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fHMM}
\alias{fHMM}
\title{fHMM: Fitting Hidden Markov Models to Financial Data}
\description{
The \{fHMM\} package provides tools for modeling financial data with
the (hierarchical) hidden Markov model.

Please see the \href{https://loelschlaeger.de/fHMM}{package website}
for more details.
}
\keyword{internal}
