% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fhmm_fit}
\alias{fhmm_fit}
\title{finite heirarchical mixture model - C++ estimation engine}
\usage{
fhmm_fit(r, n_j, d, L, K, J, mu_0, kappa_0, nu_0, sigma_0, iter_max,
  warm_up, thin, seed, chain, num_posterior_samples)
}
\arguments{
\item{r}{vector of distances associatd with different BEFs}

\item{n_j}{matrix of integers denoting the start and length of each observations associated BEF distances}

\item{d}{a 1D grid of positive real values over which the differing intensities are evaluated}

\item{L}{number of cluster mixture components}

\item{K}{number of function components}

\item{J}{number of rows in r matrix; number of groups}

\item{mu_0}{prior mean for mean parameter}

\item{kappa_0}{prior variance parameter for mean normal prior}

\item{nu_0}{prior degrees of freedom for variance Inv Chisq prior}

\item{sigma_0}{prior scale for variance Inv-Chisq distribtuion}

\item{iter_max}{number of total iterations to run the sampler for}

\item{warm_up}{number of iterations to discard as burn-in or warm_up}

\item{thin}{number of iterations to thin posterior sample draws by}

\item{seed}{seed with which to initialize random number generator}

\item{chain}{used for labeling}

\item{num_posterior_samples}{number of posterior samples kept after thinning}
}
\description{
finite heirarchical mixture model - C++ estimation engine
}
