% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_controls.R
\name{fHMM_hierarchical}
\alias{fHMM_hierarchical}
\title{Helper function that checks if the model is hierarchical}
\usage{
fHMM_hierarchical(controls)
}
\arguments{
\item{controls}{Either a \code{list} or an object of class \code{fHMM_controls}.

The \code{list} can contain the following elements, which are described
in more detail below:
\itemize{
\item \code{hierarchy}, defines an hierarchical HMM,
\item \code{states}, defines the number of states,
\item \code{sdds}, defines the state-dependent distributions,
\item \code{horizon}, defines the time horizon,
\item \code{period}, defines a flexible, periodic fine-scale time horizon,
\item \code{data}, a \code{list} of controls that define the data,
\item \code{fit}, a \code{list} of controls that define the model fitting,
\item \code{seed}, defines a seed for reproducibility,
\item \code{verbose}, defines whether to print progress messages.
}
Either none, all, or selected elements can be specified.

Unspecified parameters are set to their default values, see below.

Specifications in \code{controls} override individual specifications.}
}
\description{
Helper function that checks if the model is hierarchical
}
\keyword{internal}
