% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma2delta.R
\name{Gamma2delta}
\alias{Gamma2delta}
\title{Computes stationary distribution of transition probability matrix.}
\usage{
Gamma2delta(Gamma)
}
\arguments{
\item{Gamma}{Transition probability matrix.}
}
\value{
Stationary distribution vector.
}
\description{
Computes stationary distribution of transition probability matrix.
}
