% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma2gammasCon.R
\name{Gamma2gammasCon}
\alias{Gamma2gammasCon}
\title{Constrains non-diagonal matrix elements of transition probability matrix.}
\usage{
Gamma2gammasCon(Gamma)
}
\arguments{
\item{Gamma}{Transition probability matrix.}
}
\value{
Vector of constrained non-diagonal matrix elements (column-wise).
}
\description{
Constrains non-diagonal matrix elements of transition probability matrix.
}
\details{
Function shifts 0 and 1 non-diagonal elements by \code{1e-3}.
}
