% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasCon2Gamma.R
\name{gammasCon2Gamma}
\alias{gammasCon2Gamma}
\title{Build transition probability matrix from constrained non-diagonal elements}
\usage{
gammasCon2Gamma(gammasCon, dim)
}
\arguments{
\item{gammasCon}{Constrained non-diagonal elements of transition probability matrix}

\item{dim}{Dimension of transition probability matrix}
}
\value{
Transition probability matrix
}
\description{
Build transition probability matrix from constrained non-diagonal elements
}
