% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{gammasUncon2Gamma}
\alias{gammasUncon2Gamma}
\title{This function builds a transition probability matrix from un-constrained
non-diagonal elements \code{gammasUncon}.}
\usage{
gammasUncon2Gamma(gammasUncon, dim)
}
\arguments{
\item{gammasUncon}{A vector of un-constrained non-diagonal elements of a transition probability
matrix.}

\item{dim}{The dimension of the transition probability matrix.}
}
\value{
A transition probability matrix.
}
\description{
This function builds a transition probability matrix from un-constrained
non-diagonal elements \code{gammasUncon}.
}
\keyword{internal}
