% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_initial_estimate.R
\name{get_initial_estimate}
\alias{get_initial_estimate}
\title{Initial estimate for numerical likelihood optimization}
\usage{
get_initial_estimate(data, verbose = TRUE, initial_estimate = NULL)
}
\arguments{
\item{data}{An object of class \code{\link{fHMM_data}}.}

\item{verbose}{Set to \code{TRUE} to print progress messages.}

\item{initial_estimate}{Optionally defines an initial estimate for the numerical likelihood
optimization. Can be:
\itemize{
\item \code{NULL} (the default), in this case
\itemize{
\item applies a heuristic to calculate an initial estimate,
\item or uses the true parameter values if \code{data$controls$origin} is
\code{TRUE}.
}
\item or an object of class \code{parUncon}, for example the estimate of a
previously fitted model, i.e. the element \code{model$estimate}.
}}
}
\value{
An object of class \code{parUncon}.
}
\description{
This helper function uses a heuristic to find a good initial estimate for the
numerical optimization of the model likelihood function.
}
\keyword{internal}
