% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_loss.R
\name{green_loss.hmm}
\alias{green_loss.hmm}
\title{Lau and Green posterior loss function}
\usage{
\method{green_loss}{hmm}(object, tau = 0.5)
}
\arguments{
\item{tau}{loss function tuning parameter}
}
\description{
Lau and Green posterior loss function
}
