% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{hmm}
\alias{hmm}
\title{Create an hmm object}
\usage{
hmm(object, chains)
}
\arguments{
\item{object}{A list provided by the fhmm function}

\item{chains}{number of chains}
}
\value{
An ndp object
}
\description{
Create an hmm object
}
