% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_est.R
\name{init_est}
\alias{init_est}
\title{Initialisation}
\usage{
init_est(controls)
}
\arguments{
\item{controls}{A list of controls.}
}
\value{
A vector of parameters values in format \code{thetaUncon}.
}
\description{
Samples initial parameter values for the estimation routine.
}
