% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_number.R
\name{is_number}
\alias{is_number}
\title{Check if \code{x} is a ((non)-negative) ((non-)positive) (integer) numeric
(vector).}
\usage{
is_number(
  x,
  int = FALSE,
  neg = FALSE,
  non_neg = FALSE,
  pos = FALSE,
  non_pos = FALSE
)
}
\arguments{
\item{x}{An R object.}

\item{int}{A boolean, if \code{TRUE} checks if \code{x} is an integer.}

\item{neg}{A boolean, if \code{TRUE} checks if \code{x} is negative.}

\item{non_neg}{A boolean, if \code{TRUE} checks if \code{x} is non-negative.}

\item{pos}{A boolean, if \code{TRUE} checks if \code{x} is positive.}

\item{non-pos}{A boolean, if \code{TRUE} checks if \code{x} is non-positive.}
}
\value{
A boolean.
}
\description{
Check if \code{x} is a ((non)-negative) ((non-)positive) (integer) numeric
(vector).
}
