% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tpm}
\alias{is_tpm}
\title{Check if \code{x} is a transition probability matrix.}
\usage{
is_tpm(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A boolean.
}
\description{
Check if \code{x} is a transition probability matrix.
}
\keyword{utils}
