% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{logLik}
\alias{logLik}
\alias{logLik.fHMM_model}
\title{Log-likelihood value}
\usage{
logLik(object, ...)

\method{logLik}{fHMM_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fHMM_model}.}

\item{...}{Ignored.}
}
\value{
A numeric.
}
\description{
This function computes the log-likelihood value of an \code{fHMM_model}
object.
}
\examples{
data("dax_model_3t", package = "fHMM")
logLik(dax_model_3t)
}
