% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_all}
\alias{match_all}
\title{Match positions of two numeric vectors "as good as possible".}
\usage{
match_all(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{Another numeric vector of the same length as \code{x}.}
}
\value{
An integer vector of length \code{length(x)} with the positions of \code{y}
in \code{x}.
}
\description{
This function matches the positions of two numeric vectors.
}
\keyword{utils}
