% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_likelihood.R
\name{max_likelihood}
\alias{max_likelihood}
\title{Optimization}
\usage{
max_likelihood(data, controls)
}
\arguments{
\item{data}{A list of processed data information.}

\item{controls}{A list of controls.}
}
\value{
A list of fitted model information.
}
\description{
Maximizes the model's log-likelihood function.
}
\details{
Uses \code{nlm} for numerical optimization.
}
