% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{muUncon2muCon}
\alias{muUncon2muCon}
\title{Constrains expected values.}
\usage{
muUncon2muCon(muUncon, link)
}
\arguments{
\item{muUncon}{Vector of unconstrained expected values.}

\item{link}{Boolean, determining whether to apply the link function.}
}
\value{
Vector of constrained expected values.
}
\description{
Constrains expected values.
}
\keyword{internal}
