% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{nobs}
\alias{nobs}
\alias{nobs.fHMM_model}
\title{Number of observations}
\usage{
nobs(object, ...)

\method{nobs}{fHMM_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fHMM_model}.}

\item{...}{Ignored.}
}
\value{
An integer.
}
\description{
This function extracts the number of observations from an \code{fHMM_model}
object.
}
\examples{
data("dax_model_3t", package = "fHMM")
nobs(dax_model_3t)

}
