% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{par2parCon}
\alias{par2parCon}
\title{This function transforms an object of class \code{fHMM_parameters} into an
object of class \code{parCon}.}
\usage{
par2parCon(par, controls)
}
\arguments{
\item{par}{An object of class \code{fHMM_parameters}.}

\item{controls}{An object of class{fHMM_controls}.}
}
\value{
An object of class \code{parCon}.
}
\description{
This function transforms an object of class \code{fHMM_parameters} into an
object of class \code{parCon}.
}
\keyword{internal}
