% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{par2parUncon}
\alias{par2parUncon}
\title{This function transforms an object of class \code{fHMM_parameters} into
an object of class \code{parUncon}.}
\usage{
par2parUncon(par, controls)
}
\arguments{
\item{par}{An object of class \code{fHMM_parameters}.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
An object of class \code{parUncon}, i.e. a \code{numeric} vector of
unconstrained model parameters to be estimated.
}
\description{
This function transforms an object of class \code{fHMM_parameters} into
an object of class \code{parUncon}.
}
\keyword{internal}
