% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{parameter_labels}
\alias{parameter_labels}
\title{Create labels for estimated parameters}
\usage{
parameter_labels(controls, expected_length = NULL)
}
\arguments{
\item{controls}{[\code{fHMM_controls}]\cr
An object of class \code{fHMM_controls}.}

\item{expected_length}{[\code{NULL} | \code{integer(1)}]\cr
The expected output length. If \code{NULL} (default), this is not checked.}
}
\value{
A \code{character} vector of parameter labels.
}
\description{
This helper function creates labels for the estimated HMM parameters.
}
\keyword{internal}
