% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_names.R
\name{parameter_names}
\alias{parameter_names}
\title{Create model parameter names}
\usage{
parameter_names(controls, all)
}
\arguments{
\item{controls}{A list of controls}

\item{all}{A boolean, determining whether all (\code{all=TRUE}) or only estimated (\code{all=FALSe}) names should be produced}
}
\value{
Vector of model parameter names
}
\description{
Create model parameter names
}
