% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{parCon2par}
\alias{parCon2par}
\title{This function transforms an object of class \code{parCon} into an object
of class \code{fHMM_parameters}.}
\usage{
parCon2par(parCon, controls)
}
\arguments{
\item{parCon}{An object of class \code{parCon}.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
An object of class \code{fHMM_parameters}.
}
\description{
This function transforms an object of class \code{parCon} into an object
of class \code{fHMM_parameters}.
}
\keyword{internal}
