% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{parUncon2par}
\alias{parUncon2par}
\title{This function transforms an object of class \code{parUncon} into an
object of class \code{fHMM_parameters}.}
\usage{
parUncon2par(parUncon, controls)
}
\arguments{
\item{parUncon}{An object of class \code{parUncon}.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
An object of class \code{fHMM_parameters}.
}
\description{
This function transforms an object of class \code{parUncon} into an
object of class \code{fHMM_parameters}.
}
\keyword{internal}
