% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{parUncon2parCon}
\alias{parUncon2parCon}
\title{This function transforms an object of class \code{parUncon} into an object
of class \code{parCon}.}
\usage{
parUncon2parCon(parUncon, controls)
}
\arguments{
\item{parUncon}{An object of class \code{parUncon}.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
An object of class \code{parCon}, i.e. a vector of constrained model
parameters to be estimated.
}
\description{
This function transforms an object of class \code{parUncon} into an object
of class \code{parCon}.
}
\keyword{internal}
