% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.hmm}
\alias{plot.hmm}
\title{Plot method for hmm objects}
\usage{
\method{plot}{hmm}(x, plotfun = "cluster")
}
\arguments{
\item{x}{an hmm object}

\item{plotfun}{one of "cluster","global", or "network"}
}
\description{
The \code{plot} method for \link{hmm-objects} provides convenient
access to several types of plots useful for visualizing
hmm models using the default settings.
}
