% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_global_density.hmm}
\alias{plot_global_density.hmm}
\title{Plots global density function - for monitoring convergence}
\usage{
\method{plot_global_density}{hmm}(x, p = 0.9, r = NULL,
  transform = TRUE)
}
\arguments{
\item{x}{hmm object}

\item{p}{probability mass contained in uncertainty interval}
}
\value{
ggplot plot object
}
\description{
Plots global density function - for monitoring convergence
}
