% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pairs.hmm}
\alias{plot_pairs.hmm}
\title{plots pairwise probability clustering plot}
\usage{
\method{plot_pairs}{hmm}(x)
}
\arguments{
\item{x}{hmm object}
}
\value{
ggplot plot object
}
\description{
plots pairwise probability clustering plot
}
