% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdd.R
\name{plot_sdd}
\alias{plot_sdd}
\title{Visualization of estimated state-dependent distributions}
\usage{
plot_sdd(controls, data, fit, decoding, colors)
}
\arguments{
\item{controls}{A list of controls.}

\item{data}{A list of processed data information.}

\item{fit}{A list of fitted model information.}

\item{decoding}{A vector (in case of a HMM) or a matrix (in case of a hierarchical HMM) of decoded states.}

\item{colors}{A matrix of colors for different states.}
}
\value{
No return value. Creates file "state_dependent_distributions.pdf" in "\code{controls[["path"]]}/models/\code{controls[["id"]]}".
}
\description{
Plots the estimated state-dependent distributions.
}
