% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_traceplots.hmm}
\alias{plot_traceplots.hmm}
\title{Traceplots of various NDP-NHPP parameters}
\usage{
\method{plot_traceplots}{hmm}(x, par = "alpha")
}
\arguments{
\item{x}{hmm object}

\item{par}{character vector of parameter names, defaults to c("alpha")}
}
\value{
ggplot plot object
}
\description{
Traceplots of various NDP-NHPP parameters
}
