% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Prediction}
\usage{
predict(model, ahead, ci_level = 0.05)
}
\arguments{
\item{model}{An object of class \code{fHMM_model}.}

\item{ahead}{A positive integer.}

\item{ci_level}{The confidence level, a numeric between 0 and 1.}
}
\value{
An object of class \code{fHMM_predict}.
}
\description{
This function...
}
