% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_controls.R
\name{print.fHMM_controls}
\alias{print.fHMM_controls}
\title{Print method for \code{fHMM_controls}.}
\usage{
\method{print}{fHMM_controls}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_controls}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_controls}.
}
