% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_data.R
\name{print.fHMM_data}
\alias{print.fHMM_data}
\title{Print method for \code{fHMM_data}.}
\usage{
\method{print}{fHMM_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_data}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_data}.
}
