% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_events.R
\name{print.fHMM_events}
\alias{print.fHMM_events}
\title{Print method for \code{fHMM_events}.}
\usage{
\method{print}{fHMM_events}(x)
}
\arguments{
\item{x}{An object of class \code{fHMM_events}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_events}.
}
