% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_model.R
\name{print.fHMM_model}
\alias{print.fHMM_model}
\title{Print method for \code{fHMM_model}.}
\usage{
\method{print}{fHMM_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_model}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_model}.
}
