% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_parameters.R
\name{print.fHMM_parameters}
\alias{print.fHMM_parameters}
\title{Print method for \code{fHMM_parameters}.}
\usage{
\method{print}{fHMM_parameters}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_parameters}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_parameters}.
}
