% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_predict.R
\name{print.fHMM_predict}
\alias{print.fHMM_predict}
\title{Print method for \code{fHMM_predict}.}
\usage{
\method{print}{fHMM_predict}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_predict}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_predict}.
}
