% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_sdds.R
\name{print.fHMM_sdds}
\alias{print.fHMM_sdds}
\title{Print method for \code{fHMM_sdds}.}
\usage{
\method{print}{fHMM_sdds}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_sdds}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
This function is the print method for an object of class \code{fHMM_sdds}.
}
