% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fHMM_trouble.R
\name{print.fHMM_trouble}
\alias{print.fHMM_trouble}
\title{Print method for \code{fHMM_trouble}.}
\usage{
\method{print}{fHMM_trouble}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_trouble}.}

\item{...}{Ignored.}
}
\value{
No return value.
}
\description{
This function is the print method for an object of class \code{fHMM_trouble}.
}
