% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{print.hmm}
\alias{print.hmm}
\title{Print method for hmm objects}
\usage{
\method{print}{hmm}(x, digits = 2, ...)
}
\arguments{
\item{digits}{Number of digits to use for formatting numbers.}

\item{...}{Ignored.}
}
\value{
Returns \code{x}, invisibly.  #' @details 
\subsection{Point estimates}{
Point estimates are medians computed from simulations.
For models fit using MCMC (\code{"sampling"}) the posterior
sample is used.  The point estimates reported are the same as the values
returned by \code{\link[=coef.stapreg]{coef}}.
}
\subsection{Uncertainty estimates (MAD_SD)}{
The standard deviations reported (labeled \code{MAD_SD} in the print output)
are computed from the same set of draws described above and are proportional
to the median absolute deviation (\code{\link[stats]{mad}}) from the median.
Compared to the raw posterior standard deviation, the MAD_SD will be
more robust for long-tailed distributions. 
}
}
\description{
The \code{print} method for hmm objects displays a compact summary of the
fitted model. See the \strong{Details} section below for descriptions of the
different components of the printed output. For additional summary statistics
and diagnostics use the \code{\link[=summary.hmm]{summary}} method.
}
\seealso{
\code{\link{summary.hmm}}, \code{\link{stapreg-methods}}
}
