% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.fHMM_data.R
\name{print.summary.fHMM_data}
\alias{print.summary.fHMM_data}
\title{Print method for the summary of \code{fHMM_data}.}
\usage{
\method{print}{summary.fHMM_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.fHMM_data}.}

\item{...}{Ignored.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print method for the summary of \code{fHMM_data}.
}
