% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Data processing}
\usage{
process_data(controls, sim_par)
}
\arguments{
\item{controls}{A list of controls.}

\item{sim_par}{A vector of model parameters for simulation.}
}
\value{
A list of processed data information and on-screen information.
}
\description{
Calls functions for processing or simulating data.
}
