% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress.R
\name{progress}
\alias{progress}
\title{Print progess.}
\usage{
progress(run, total_runs, start_time, message = NULL)
}
\arguments{
\item{run}{The current iteration. Set \code{run = 0} for an initial message.}

\item{total_runs}{The total number of iterations.}

\item{start_time}{The time at which the loop started.}

\item{message}{A character, a message printed in front of the progress bar.
The initial message.}
}
\value{
No return value, only on-screen information.
}
\description{
This function prints the progress of some loop.
}
\examples{
start_time = Sys.time()
total_runs = 10
progress(run = 0, total_runs = total_runs, start_time = start_time,
         message = "Progress:")
for(run in 1:total_runs){
  Sys.sleep(0.5)
  progress(run = run, total_runs = total_runs, start_time = start_time,
           message = "Progress:")
}
}
