% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_residuals.R
\name{pseudo_residuals}
\alias{pseudo_residuals}
\title{Pseudo-residuals}
\usage{
pseudo_residuals(controls, data, fit, decoding)
}
\arguments{
\item{controls}{A list of controls.}

\item{data}{A list of processed data information.}

\item{fit}{A list of fitted model information.}

\item{decoding}{A vector (in case of a HMM) or a matrix (in case of a hierarchical HMM) of decoded states.}
}
\value{
No return value. Creates files "pseudo_residuals.pdf" and "pseudos.rds" in "\code{controls[["path"]]}/models/\code{controls[["id"]]}".
}
\description{
Computes and visualizes pseudo-residuals.
}
