% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_model.R
\name{residuals.fHMM_model}
\alias{residuals.fHMM_model}
\title{Residuals}
\usage{
\method{residuals}{fHMM_model}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{fHMM_model}}.}

\item{...}{Ignored.}
}
\value{
A vector (or a matrix, in case of an hierarchical HMM) with (pseudo-)
residuals for each observation.
}
\description{
This function extracts the computed (pseudo-) residuals of
an \code{\link{fHMM_model}} object.
}
\examples{
compute_residuals(dax_model_3t)
res <- residuals(dax_model_3t)
head(res)

}
