% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_tpm}
\alias{sample_tpm}
\title{Sample transition probability matrices}
\usage{
sample_tpm(dim, state_persistent = TRUE)
}
\arguments{
\item{dim}{An \code{integer}, the matrix dimension.}

\item{state_persistent}{Set to \code{TRUE} (default) to put more probability on the diagonal.}
}
\value{
A transition probability \code{matrix}.
}
\description{
This function returns a random, squared matrix of dimension \code{dim}
that fulfills the properties of a transition probability matrix.
}
\examples{
\dontrun{
sample_tpm(dim = 3)
}
}
\keyword{utils}
