% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{set_default_arguments}
\alias{set_default_arguments}
\title{Set default arguments to a \code{function}}
\usage{
set_default_arguments(f, default_arguments = list())
}
\arguments{
\item{f}{A \code{function}.}

\item{default_arguments}{A named \code{list} of default arguments to \code{f}.}
}
\value{
The input \code{f} with defined default arguments.
}
\description{
This helper function defines default arguments for a \code{function}.
}
\examples{
\dontrun{
f <- function(x) x
f <- set_default_arguments(f, list("x" = 1))
f()
}

}
\keyword{internal}
