% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mfrow.R
\name{set_mfrow}
\alias{set_mfrow}
\title{Balancing visualization of multiple figures.}
\usage{
set_mfrow(n)
}
\arguments{
\item{n}{The total number of figures.}
}
\value{
A vector of the form \code{c(nr,nc)}. If \code{par(mfrow = c(nr,nc))},
subsequent figures will be drawn in an \code{nr} x \code{nc} array on the
current device by rows.
}
\description{
This function finds a balanced setting for \code{par(mfrow)}.
}
