% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaCon2sigmaUncon.R
\name{sigmaCon2sigmaUncon}
\alias{sigmaCon2sigmaUncon}
\title{Unconstrain standard deviations}
\usage{
sigmaCon2sigmaUncon(sigmaCon)
}
\arguments{
\item{sigmaCon}{Constrained standard deviations}
}
\value{
Unconstrained standard deviations
}
\description{
Unconstrain standard deviations
}
