% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{sigmaUncon2sigmaCon}
\alias{sigmaUncon2sigmaCon}
\title{This function constrains the un-constrained standard deviations
\code{sigmaUncon}.}
\usage{
sigmaUncon2sigmaCon(sigmaUncon)
}
\arguments{
\item{sigmaUncon}{A vector of un-constrained standard deviations.}
}
\value{
A vector of constrained standard deviations.
}
\description{
This function constrains the un-constrained standard deviations
\code{sigmaUncon}.
}
\keyword{internal}
