% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{sim_model_4lnorm}
\alias{sim_model_4lnorm}
\title{Simulated 4-state HMM with log-normal distributions}
\format{
An object of class \code{fHMM_model}.
}
\usage{
data("sim_model_4lnorm")
}
\description{
A pre-computed 4-state HMM with state-dependent log-normal distributions
on \code{1000} simulated observations.
}
\details{
The model was estimated via:
\preformatted{
controls <- list(
  states  = 4,
  sdds    = "lnorm",
  horizon = 1000,
  fit     = list(runs = 50)
)
controls <- set_controls(controls)
data <- prepare_data(controls, seed = 1)
sim_model_4lnorm <- fit_model(data, seed = 1, verbose = TRUE)
}
}
\keyword{model}
