% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{spx}
\alias{spx}
\title{Standard & Poor’s 500 (S&P 500) index data}
\format{
A \code{data.frame} with 23864 rows and the following 7 columns:
\itemize{
\item \code{Date}: The date.
\item \code{Open}: Opening price.
\item \code{High}: Highest price.
\item \code{Low}: Lowest price.
\item \code{Close}: Close price adjusted for splits.
\item \code{Adj.Close}: Close price adjusted for dividends and splits.
\item \code{Volume}: Trade volume.
}
}
\usage{
spx
}
\description{
Standard & Poor’s 500 (S&P 500) index data from 1928 to 2022 from Yahoo
Finance.
}
\details{
The data was obtained via:
\preformatted{
spx <- download_data(
  symbol = "^GSPC",    # S&P 500 identifier on Yahoo Finance
  from = "1928-01-01", # first observation
  to = "2022-12-31"    # last observation
)
}
}
\keyword{data}
