% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{spx_unemp_model}
\alias{spx_unemp_model}
\title{Unemployment rate and S&P 500 hierarchical HMM}
\format{
An object of class \code{\link{fHMM_model}}.
}
\usage{
data("spx_unemp_model")
}
\description{
A pre-computed HHMM with ... TODO
}
\details{
The model was estimated via:
\preformatted{
controls <- list(
  hierarchy = TRUE,
  states    = c(2, 2),
  sdds      = c("gamma", "t"),
  period    = "m",
  data      = list(
    file       = list(unemp, spx),
    date_column = c("date", "Date"),
    data_column = c("rate", "Close"),
    from       = "2015-01-01",
    to         = "2022-12-31",
    logreturns = c(FALSE, TRUE)),
  fit = list(runs = 4)
)
controls <- set_controls(controls)
spx_unemp_data <- prepare_data(controls)
spx_unemp_model <- fit_model(spx_unemp_data)
spx_unemp_model <- decode_states(spx_unemp_model)
spx_unemp_model <- compute_residuals(spx_unemp_model)
}
}
\keyword{model}
