% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stationary_distribution}
\alias{stationary_distribution}
\title{Compute stationary distribution}
\usage{
stationary_distribution(Gamma)
}
\arguments{
\item{Gamma}{A transition probability \code{matrix}.}
}
\value{
A \code{numeric} vector, the stationary distribution.
}
\description{
This function computes the stationary distribution of a transition
probability matrix \code{Gamma}.
}
\details{
If the stationary distribution vector cannot be computed, it is set to the
discrete uniform distribution over the states.
}
\examples{
\dontrun{
Gamma <- matrix(c(0.5, 0.3, 0.5, 0.7), 2, 2)
stationary_distribution(Gamma)
}
}
\keyword{utils}
