% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fHMM_data.R
\name{summary.fHMM_data}
\alias{summary.fHMM_data}
\title{Summary method for \code{fHMM_data}.}
\usage{
\method{summary}{fHMM_data}(object)
}
\arguments{
\item{object}{An object of class \code{fHMM_data}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.fHMM_data}.
}
\description{
Summary method for \code{fHMM_data}.
}
