% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fHMM_model.R
\name{summary.fHMM_model}
\alias{summary.fHMM_model}
\title{Summary method for \code{fHMM_model}.}
\usage{
\method{summary}{fHMM_model}(object)
}
\arguments{
\item{object}{An object of class \code{fHMM_model}.}

\item{...}{Ignored.}
}
\value{
An object of class \code{summary.fHMM_model}.
}
\description{
Summary method for \code{fHMM_model}.
}
