% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaCon2thetaList.R
\name{thetaCon2thetaList}
\alias{thetaCon2thetaList}
\title{Bring contrained model parameters from vector form to list form}
\usage{
thetaCon2thetaList(thetaCon, controls)
}
\arguments{
\item{thetaCon}{Constrained model parameters in vector form}

\item{controls}{A list of controls}
}
\value{
Constrained model parameters in list form
}
\description{
Bring contrained model parameters from vector form to list form
}
