% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaCon2thetaUncon.R
\name{thetaCon2thetaUncon}
\alias{thetaCon2thetaUncon}
\title{Unconstrains model parameters.}
\usage{
thetaCon2thetaUncon(thetaCon, controls)
}
\arguments{
\item{thetaCon}{Constrained model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
Unconstrained model parameters in vector form.
}
\description{
Unconstrains model parameters.
}
