% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaList2thetaCon.R
\name{thetaList2thetaCon}
\alias{thetaList2thetaCon}
\title{Bring constrained model parameters from list form to vector form}
\usage{
thetaList2thetaCon(thetaList, controls)
}
\arguments{
\item{thetaList}{Constrained model parameters in list form}

\item{controls}{A list of controls}
}
\value{
Constrained model parameters in vector form
}
\description{
Bring constrained model parameters from list form to vector form
}
