% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaUncon2thetaCon.R
\name{thetaUncon2thetaCon}
\alias{thetaUncon2thetaCon}
\title{Constrains model parameters.}
\usage{
thetaUncon2thetaCon(thetaUncon, controls)
}
\arguments{
\item{thetaUncon}{Unconstrained model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
Constrained model parameters in vector form.
}
\description{
Constrains model parameters.
}
