% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaUncon2thetaList.R
\name{thetaUncon2thetaList}
\alias{thetaUncon2thetaList}
\title{Bring uncontrained model parameters in list form}
\usage{
thetaUncon2thetaList(thetaUncon, controls)
}
\arguments{
\item{thetaUncon}{Unconstrained model parameters in vector form}

\item{controls}{A list of controls}
}
\value{
Constrained model parameters in list form
}
\description{
Bring uncontrained model parameters in list form
}
