% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-and-summary.R
\name{to_latex.hmm}
\alias{to_latex.hmm}
\title{converts summary output to latex table}
\usage{
\method{to_latex}{hmm}(object, digits = 1, caption = "",
  par = c("alpha"))
}
\description{
converts summary output to latex table
}
