% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trouble.R
\name{trouble}
\alias{trouble}
\title{Debugging.}
\usage{
trouble(code)
}
\arguments{
\item{code}{A character, the exception code.}
}
\value{
An object of class \code{fHMM_trouble}, which is a list containing the
following elements:
\itemize{
\item \code{code}: The input \code{code}.
\item \code{response}: The output message.
\item \code{debugging}: Suggestions for debugging.
}
}
\description{
Provides suggestions for debugging for a given exception code.
}
\examples{
trouble("S.1")
}
