% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_controls.R
\name{validate_controls}
\alias{validate_controls}
\title{Validate control settings}
\usage{
validate_controls(controls)
}
\arguments{
\item{controls}{Either a \code{list} or an object of class \code{fHMM_controls}.}
}
\value{
The checked input \code{controls}.
}
\description{
This helper function validates control settings.
}
\keyword{internal}
