% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{vw}
\alias{vw}
\title{VW data}
\format{
A \code{data.frame} with the columns:
TODO
}
\usage{
vw
}
\description{
VW data from ... to ... scraped from ...
TODO
}
\details{
The data was obtained via ...
TODO
}
\keyword{data}
