% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_viterbi.R
\name{apply_viterbi}
\alias{apply_viterbi}
\title{Apply the Viterbi algorithm for state decoding}
\usage{
apply_viterbi(data, fit, controls)
}
\arguments{
\item{data}{A list of processed data information}

\item{fit}{A fitted model}

\item{controls}{A list of controls}
}
\value{
A vector (in case of HMM) or a matrix (in case of HHMM) of decoded states
}
\description{
Apply the Viterbi algorithm for state decoding
}
