% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_controls.R
\name{check_controls}
\alias{check_controls}
\title{Check specification of controls}
\usage{
check_controls(controls)
}
\arguments{
\item{controls}{A list of controls}
}
\value{
Checked \code{controls}
}
\description{
Check specification of controls
}
