% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date.R
\name{check_date}
\alias{check_date}
\title{Check date format "YYYY-MM-DD".}
\usage{
check_date(date)
}
\arguments{
\item{date}{A character, specifying a date in format "YYYY-MM-DD".}
}
\value{
\code{as.Date(date)} if \code{date} has the format "YYYY-MM-DD".
Otherwise, the function throws an error.
}
\description{
This function checks if the input \code{date} has the format "YYYY-MM-DD".
}
