% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_date}
\alias{check_date}
\title{Check date format}
\usage{
check_date(date)
}
\arguments{
\item{date}{A \code{character}, specifying a date in format \code{"YYYY-MM-DD"}.}
}
\value{
\code{as.Date(date)} if \code{date} has the format \code{"YYYY-MM-DD"}.
Otherwise, the function throws an error.
}
\description{
This function checks if the input \code{date} has the format
\code{"YYYY-MM-DD"}.
}
\examples{
\dontrun{
check_date(date = "2000-01-01")
}
}
\keyword{internal}
\keyword{utils}
