% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_decoding.R
\name{check_decoding}
\alias{check_decoding}
\title{Check and save decoded states}
\usage{
check_decoding(decoding, data, controls)
}
\arguments{
\item{decoding}{A vector (in case of HMM) or a matrix (in case of HHMM) of decoded states}

\item{data}{A list of processed data information}

\item{controls}{A list of controls}
}
\value{
No return value, called for side effects
}
\description{
Check and save decoded states
}
