% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_estimation.R
\name{check_estimation}
\alias{check_estimation}
\title{Check and save estimates}
\usage{
check_estimation(mods, llks, data, hessian, controls)
}
\arguments{
\item{mods}{A list of fitted models in the different estimation runs}

\item{llks}{A vector of log-likelihood values of accepted \code{mods}}

\item{data}{A list of processed data information}

\item{hessian}{The Hessian matrix}

\item{controls}{A list of controls}
}
\value{
A fitted model
}
\description{
Check and save estimates
}
