% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_saving.R
\name{check_saving}
\alias{check_saving}
\title{Saving check}
\usage{
check_saving(object = NULL, name = NULL, filetype, controls)
}
\arguments{
\item{object}{An .rds-object to be saved, default \code{NULL}.}

\item{name}{A character, the name of the object to be saved, default \code{NULL}.}

\item{filetype}{A character, the filetype of the object to be saved.}

\item{controls}{A list of controls.}
}
\value{
A boolean, determining whether saving is possible or not. If \code{filetype="rds"}, \code{object} is saved.
}
\description{
Checks for overwriting and saves model results.
}
