% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_saving.R
\name{check_saving}
\alias{check_saving}
\title{Check for overwriting and save model results}
\usage{
check_saving(object = NULL, name = NULL, filetype, controls)
}
\arguments{
\item{object}{An object to be saved, default \code{NULL}}

\item{name}{The name of the object to be saved, default \code{NULL}}

\item{filetype}{The filetype of the object to be saved}

\item{controls}{A list of controls}
}
\value{
A boolean, determining whether saving is possible or not
}
\description{
Check for overwriting and save model results
}
