% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{coef.fHMM_model}
\alias{coef.fHMM_model}
\title{Model coefficients}
\usage{
\method{coef}{fHMM_model}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An object of class \code{fHMM_model}.}

\item{alpha}{The alpha level for the confidence interval, a numeric between 0 and 1.
Per default, \code{alpha = 0.05}, which computes a 95\% confidence interval.}

\item{...}{Ignored.}
}
\value{
A data frame.
}
\description{
This function returns the estimated model coefficients and an \code{alpha}
confidence interval.
}
