% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare multiple models}
\usage{
compare_models(...)
}
\arguments{
\item{...}{A list of one or more objects of class \code{\link{fHMM_model}}.}
}
\value{
A \code{data.frame} with models in rows and comparison criteria in columns.
}
\description{
This function performs model comparison by comparing multiple
\code{\link{fHMM_model}} objects with respect to
\itemize{
\item the number of model parameters,
\item the log-likelihood value,
\item the AIC value,
\item the BIC value.
}
}
\examples{
### 3-state HMM with t-distributions is preferred over 2-state HMM with
### normal distributions for the DAX data based on AIC and BIC
compare_models(dax_model_2n, dax_model_3t)

}
