% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Comparing multiple \code{fHMM_model}-objects}
\usage{
compare_models(...)
}
\arguments{
\item{...}{A list of one or more objects of class \code{fHMM_model}.}
}
\value{
A data frame with models in rows and comparison criteria in columns.
}
\description{
This function compares multiple \code{fHMM_model} with respect to
\itemize{
\item the number of model parameters,
\item the log-likelihood value,
\item the AIC value,
\item the BIC value.
}
}
\examples{
data(dax_model_3t)
compare_models(dax_model_3t)
}
