% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci}
\alias{compute_ci}
\title{Confidence intervals}
\usage{
compute_ci(fit, controls)
}
\arguments{
\item{fit}{A list of fitted model information.}

\item{controls}{A list of controls.}
}
\value{
A list containing the following elements:
\item{lb_ci_level}{lower bound of the intervals}
\item{estimate}{estimates}
\item{ub_ci_level}{upper bound of the intervals}
where \code{ci_level} is set in \code{controls}.
}
\description{
Computes confidence intervals for the estimates.
}
