% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci}
\alias{compute_ci}
\title{Confidence intervals}
\usage{
compute_ci(x, ci_level = 0.05)
}
\arguments{
\item{x}{An object of class \code{fHMM_model}.}

\item{ci_level}{The confidence level, a numeric between 0 and 1.}
}
\value{
A list containing the following elements:
\itemize{
\item \code{lb}: lower bound of confidence
\item \code{estimate}: point estimate
\item \code{ub}: upper bound of confidence
}
}
\description{
Computes confidence intervals for the estimates.
}
\keyword{internal}
