% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci}
\alias{compute_ci}
\title{Compute confidence intervals for the estimates}
\usage{
compute_ci(fit, controls)
}
\arguments{
\item{fit}{A fitted model}

\item{controls}{A list of controls}
}
\value{
A list containing the following elements:
\item{lb_alpha}{the lower bound of the intervals}
\item{estimate}{the estimates}
\item{ub_limit}{the upper bound of the intervals}
}
\description{
Compute confidence intervals for the estimates
}
