% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ci.R
\name{compute_ci}
\alias{compute_ci}
\title{Compute confidence intervals}
\usage{
compute_ci(x, alpha = 0.05)
}
\arguments{
\item{x}{An object of class \code{\link{fHMM_model}}.}

\item{alpha}{A \code{numeric} between 0 and 1, the alpha level for the confidence interval.
By default, \code{alpha = 0.05}, which computes a 95\% confidence interval.}
}
\value{
A \code{list} containing the following elements:
\itemize{
\item \code{lb}: lower bound of confidence
\item \code{estimate}: point estimate
\item \code{ub}: upper bound of confidence
}
}
\description{
This helper function computes confidence intervals for the estimates of an
\code{\link{fHMM_model}} object using the inverse Fisher information.
}
\keyword{internal}
