% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fs.R
\name{compute_fs}
\alias{compute_fs}
\title{Compute (flexible) fine-scale chunk lengths}
\usage{
compute_fs(fs_time_horizon, T = NULL, fs_dates = NULL)
}
\arguments{
\item{fs_time_horizon}{Setting the fine-scale dimension, either a numeric or one of "w","m","q","y"}

\item{T}{The dimension of the coarse-scale process, default \code{NULL}}

\item{fs_dates}{A vector of dates of empirical fine-scale observations, default \code{NULL}}
}
\value{
Vector of fine-scale chunk sizes
}
\description{
Compute (flexible) fine-scale chunk lengths
}
