% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_residuals.R
\name{compute_residuals}
\alias{compute_residuals}
\title{Compute (pseudo-) residuals}
\usage{
compute_residuals(x, verbose = TRUE)
}
\arguments{
\item{x}{[\code{fHMM_model}]\cr
An object of class \code{\link{fHMM_model}}.}

\item{verbose}{[\code{logical(1)}]\cr
Set to \code{TRUE} (default) to print progress messages.}
}
\value{
An object of class \code{\link{fHMM_model}} with residuals included.
}
\description{
This function computes (pseudo-) residuals of an \code{\link{fHMM_model}}
object.
}
\examples{
compute_residuals(dax_model_3t)
summary(residuals(dax_model_3t))
}
