% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_residuals.R
\name{compute_residuals}
\alias{compute_residuals}
\title{Compute (pseudo-) residuals.}
\usage{
compute_residuals(x)
}
\arguments{
\item{x}{An object of class \code{RprobitB_model}.}
}
\value{
An object of class \code{RprobitB_model}.
}
\description{
This function computes (pseudo-) residuals of the estimated model.
}
