% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm.R
\name{compute_T_star}
\alias{compute_T_star}
\title{Compute lengths of fine-scale data chunks}
\usage{
compute_T_star(
  controls = list(),
  hierarchy = TRUE,
  horizon = if (!hierarchy) 100 else c(100, 30),
  period = if (hierarchy && is.na(horizon[2])) "m" else NA,
  dates = NULL,
  seed = NULL
)
}
\arguments{
\item{controls}{Either a \code{list} or an object of class \code{fHMM_controls}.

The \code{list} can contain the following elements, which are described
in more detail below:
\itemize{
\item \code{hierarchy}, defines an hierarchical HMM,
\item \code{states}, defines the number of states,
\item \code{sdds}, defines the state-dependent distributions,
\item \code{horizon}, defines the time horizon,
\item \code{period}, defines a flexible, periodic fine-scale time horizon,
\item \code{data}, a \code{list} of controls that define the data,
\item \code{fit}, a \code{list} of controls that define the model fitting,
\item \code{seed}, defines a seed for reproducibility,
\item \code{verbose}, defines whether to print progress messages.
}
Either none, all, or selected elements can be specified.

Unspecified parameters are set to their default values, see below.

Specifications in \code{controls} override individual specifications.}

\item{hierarchy}{A \code{logical}, set to \code{TRUE} for an hierarchical HMM.

If \code{hierarchy = TRUE}, some of the other controls must be specified for
the coarse-scale and the fine-scale layer.

By default, \code{hierarchy = FALSE}.}

\item{horizon}{A \code{numeric}, specifying the length of the time horizon.

If \code{hierarchy = TRUE}, \code{horizon} must be a \code{vector} of length
2. The first entry corresponds to the coarse-scale layer, while the second
entry corresponds to the fine-scale layer.

By default, \code{horizon = 100} if \code{hierarchy = FALSE} and
\code{horizon = c(100, 30)} if \code{hierarchy = TRUE}.

If \code{data} is specified (i.e., not \code{NA}), the first entry of
\code{horizon} is ignored and the (coarse-scale) time horizon is defined by
available data.}

\item{period}{Only relevant if \code{hierarchy = TRUE} and \code{horizon[2] = NA}.

In this case, a \code{character} which specifies a flexible, periodic
fine-scale time horizon and can be one of
\itemize{
\item \code{"w"} for a week,
\item \code{"m"} for a month,
\item \code{"q"} for a quarter,
\item \code{"y"} for a year.
}

By default, \code{period = "m"} if \code{hierarchy = TRUE} and
\code{horizon[2] = NA}, and \code{NA} else.}

\item{dates}{A \code{character} vector of dates of empirical fine-scale data (if any).
By default, \code{dates = NULL}.}

\item{seed}{An \code{integer}, a seed for reproducibility, passed on to
\code{\link{set.seed}}.

By default, \code{seed = NULL} (i.e., no seed).}
}
\value{
An \code{integer} vector of fine-scale chunk lengths.
}
\description{
This helper function computes the lengths of the fine-scale data chunks.
This is only relevant in the hierarchical case.
}
\keyword{internal}
