% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_visuals.R
\name{create_visuals}
\alias{create_visuals}
\title{Visualize model results}
\usage{
create_visuals(data, fit, decoding, controls, events = NULL)
}
\arguments{
\item{data}{A list of processed data information}

\item{fit}{A fitted model}

\item{decoding}{A vector (in case of HMM) or a matrix (in case of HHMM) of decoded states}

\item{controls}{A list of controls}

\item{events}{A list of events, default \code{NULL}}
}
\value{
No return value, called for side effects.
}
\description{
Visualize model results
}
