% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_states.R
\name{decode_states}
\alias{decode_states}
\title{Decode the underlying hidden state sequence.}
\usage{
decode_states(x)
}
\arguments{
\item{x}{An object of class \code{fHMM_model}.}
}
\value{
An object of class \code{fHMM_model}.
}
\description{
This function decodes the (most likely) underlying hidden state sequence by
applying the Viterbi algorithm
\url{https://en.wikipedia.org/wiki/Viterbi_algorithm}.
}
