% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_states.R
\name{decode_states}
\alias{decode_states}
\title{Decoding the underlying hidden state sequence}
\usage{
decode_states(x, verbose = TRUE)
}
\arguments{
\item{x}{An object of class \code{fHMM_model}.}

\item{verbose}{Set to \code{TRUE} to print progress messages.}
}
\value{
An object of class \code{fHMM_model} with decoded state sequence included.
}
\description{
This function decodes the (most likely) underlying hidden state sequence by
applying the Viterbi algorithm.
}
\examples{
data(dax_model_3t)
decode_states(dax_model_3t)
}
\references{
\url{https://en.wikipedia.org/wiki/Viterbi_algorithm}
}
