% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfCon2dfUncon.R
\name{dfCon2dfUncon}
\alias{dfCon2dfUncon}
\title{Unconstrain degrees of freedom}
\usage{
dfCon2dfUncon(dfCon)
}
\arguments{
\item{dfCon}{Constrained degrees of freedom}
}
\value{
Unconstrained degrees of freedom
}
\description{
Unconstrain degrees of freedom
}
