% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{dfCon2dfUncon}
\alias{dfCon2dfUncon}
\title{This function un-constrains the constrained degrees of freedom \code{dfCon}.}
\usage{
dfCon2dfUncon(dfCon)
}
\arguments{
\item{dfCon}{A vector of constrained degrees of freedom.}
}
\value{
A vector of un-constrained degrees of freedom.
}
\description{
This function un-constrains the constrained degrees of freedom \code{dfCon}.
}
\keyword{internal}
