% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{dfUncon2dfCon}
\alias{dfUncon2dfCon}
\title{Constrains degrees of freedom.}
\usage{
dfUncon2dfCon(dfUncon)
}
\arguments{
\item{dfUncon}{Numeric, unconstrained degrees of freedom.}
}
\value{
Numeric, constrained degrees of freedom.
}
\description{
Constrains degrees of freedom.
}
\keyword{internal}
