% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfUncon2dfCon.R
\name{dfUncon2dfCon}
\alias{dfUncon2dfCon}
\title{Constrain degrees of freedom}
\usage{
dfUncon2dfCon(dfUncon)
}
\arguments{
\item{dfUncon}{Unconstrained degrees of freedom}
}
\value{
Constrained degrees of freedom
}
\description{
Constrain degrees of freedom
}
