% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download financial data from Yahoo Finance}
\usage{
download_data(
  symbol,
  from = "1902-01-01",
  to = Sys.Date(),
  file = NULL,
  columns = c("Date", "Open", "High", "Low", "Close", "Adj.Close", "Volume"),
  fill_dates = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{symbol}{A \code{character}, the stock's symbol. It must match the identifier on
\url{https://finance.yahoo.com/}.}

\item{from}{A \code{character} in the format \code{"YYYY-MM-DD"}, setting the lower
data bound. Must not be earlier than \code{"1902-01-01"} (default).}

\item{to}{A \code{character} in the format \code{"YYYY-MM-DD"}, setting the upper
data bound. Default is the current date \code{Sys.date()}.}

\item{file}{Either
\itemize{
\item \code{NULL} (default) to return the data as a \code{data.frame},
\item or a \code{character}, the path where the data is saved as a .csv-file.
}}

\item{columns}{A \code{character}}

\item{fill_dates}{Set to \code{TRUE} to fill missing dates (e.g., days at which the stock
market is closed )with \code{NA}s.}

\item{verbose}{Set to \code{TRUE} (default) to print information about download success.}
}
\value{
A \code{data.frame} if \code{file = NULL}.
}
\description{
This function downloads stock data from \url{https://finance.yahoo.com/}.
}
\details{
Yahoo Finance provides the following historical daily data for a stock or
an index:
\itemize{
\item \code{Date}: The date.
\item \code{Open}: Opening price.
\item \code{High}: Highest price.
\item \code{Low}: Lowest price.
\item \code{Close}: Close price adjusted for splits.
\item \code{Adj.Close}: Close price adjusted for dividends and splits.
\item \code{Volume}: Trade volume.
}
}
\examples{
### download 21st century closing prices of the DAX data
data <- download_data(
  symbol = "^GDAXI", from = "2000-01-01", columns = c("Date", "Close"),
  fill_dates = TRUE
)
head(data)

}
