% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exception.R
\name{exception}
\alias{exception}
\title{Debugging}
\usage{
exception(code)
}
\arguments{
\item{code}{A character, the exception code.}
}
\value{
A list containing the following elements:
\item{code}{exception code}
\item{response}{message}
\item{debugging}{suggestions for debugging}
}
\description{
Provides suggestions for debugging for a given exception code.
}
\examples{
exception("S.1")
}
