% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM-package.R
\docType{package}
\name{fHMM-package}
\alias{fHMM}
\alias{fHMM-package}
\title{fHMM: Fitting Hidden Markov Models to Financial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fitting (hierarchical) hidden Markov models to financial data via maximum likelihood estimation. The package offers different state-dependent distributions (including normal, log-normal, t, Gamma, and Poisson), state decoding, model selection, model checking, and prediction. For a reference, see Oelschläger, L. and Adam, T. "Detecting bearish and bullish markets in financial time series using hierarchical hidden Markov models" (2021, Statistical Modelling).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://loelschlaeger.de/fHMM/}
  \item \url{https://github.com/loelschlaeger/fHMM}
  \item Report bugs at \url{https://github.com/loelschlaeger/fHMM/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Authors:
\itemize{
  \item Timo Adam \email{ta59@st-andrews.ac.uk} (\href{https://orcid.org/0000-0001-9079-3259}{ORCID})
  \item Rouven Michels \email{r.michels@uni-bielefeld.de} (\href{https://orcid.org/0000-0002-5433-6197}{ORCID})
}

}
\keyword{internal}
