% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{fHMM}
\alias{fHMM}
\title{fHMM: Fitting Hidden Markov Models to Financial Data}
\description{
This package provides tools for modeling financial data via
the (hierarchical) hidden Markov model.
}
\keyword{internal}
