% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM.R
\docType{package}
\name{fHMM}
\alias{fHMM}
\title{fHMM: A package for fitting hidden Markov models to financial data}
\description{
This package provides tools for downloading and modeling financial data via
the (hierarchical) hidden Markov model.
}
\keyword{internal}
