% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM-package.R
\docType{package}
\name{fHMM}
\alias{fHMM}
\title{fHMM: Fitting Hidden Markov Models to Financial Data}
\description{
The \{fHMM\} package provides tools for modeling financial data with
the (hierarchical) hidden Markov model.

The package offers different state-dependent distributions (including
normal, log-normal, t, Gamma, and Poisson), state decoding, model selection,
model checking, and prediction.

Please see the \href{https://loelschlaeger.de/fHMM}{package website}
for more details.
}
\references{
Oelschläger, L. and Adam, T. "Detecting bearish and bullish markets in
financial time series using  hierarchical hidden Markov models"
(2021, Statistical Modelling)
}
\author{
\itemize{
\item Lennart Oelschläger, \email{oelschlaeger.lennart@gmail.com}
\item Timo Adam, \email{ta59@st-andrews.ac.uk}
\item Rouven Michels, \email{r.michels@uni-bielefeld.de}
}
}
\keyword{internal}
