% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_colors.R
\name{fHMM_colors}
\alias{fHMM_colors}
\title{Set color scheme for visualizations in the fHMM package.}
\usage{
fHMM_colors(controls, colors = NULL)
}
\arguments{
\item{controls}{An object of class \code{fHMM_controls}.}

\item{colors}{Either \code{NULL} or a a character vector of color names or hexadecimal RGB
triplets.}
}
\value{
An object of class \code{fHMM_colors}.
}
\description{
This function defines a color scheme for visualizations in the fHMM package.
}
\keyword{s3}
