% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_events.R
\name{fHMM_events}
\alias{fHMM_events}
\alias{print.fHMM_events}
\title{Checking events}
\usage{
fHMM_events(events)

\method{print}{fHMM_events}(x, ...)
}
\arguments{
\item{events}{[\code{list()}]\cr
A \code{list} of two elements.
\itemize{
\item The first element is named \code{"dates"} and contains a \code{character}
vector in format \code{"YYYY-MM-DD"}.
\item The second element is named \code{"labels"} and is a \code{character}
vector of the same length as \code{"dates"}.
}}

\item{x}{[\code{fHMM_events}]\cr
An object of class \code{fHMM_events}.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{fHMM_events}.
}
\description{
This function checks the input \code{events}.
}
\examples{
events <- list(
  dates = c("2001-09-11", "2008-09-15", "2020-01-27"),
  labels = c(
    "9/11 terrorist attack", "Bankruptcy Lehman Brothers",
    "First COVID-19 case Germany"
  )
)
events <- fHMM_events(events)

}
