% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_events.R
\name{fHMM_events}
\alias{fHMM_events}
\title{Check events.}
\usage{
fHMM_events(events)
}
\arguments{
\item{events}{A list of two elements.
The first element is named \code{"dates"} and contains characters in format
"YYYY-MM-DD".
The second element is named \code{"labels"} and is a character vector of the
same length as \code{"dates"}.}
}
\value{
An object of class \code{fHMM_data}.
}
\description{
This function checks the input \code{events}.
}
\keyword{s3}
