% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_model.R
\name{fHMM_model}
\alias{fHMM_model}
\title{Constructor of an \code{fHMM_model} object}
\usage{
fHMM_model(
  data,
  estimate,
  nlm_output,
  estimation_time,
  ll,
  lls,
  gradient,
  hessian,
  decoding
)
}
\arguments{
\item{data}{TODO}

\item{estimate}{TODO}

\item{nlm_output}{TODO}

\item{estimation_time}{TODO}

\item{ll}{TODO}

\item{lls}{TODO}

\item{gradient}{TODO}

\item{hessian}{TODO}

\item{decoding}{TODO}
}
\value{
An object of class \code{\link{fHMM_model}}, which is a \code{list} containing
the following elements:
TODO
}
\description{
This function constructs an object of class \code{fHMM_model}, which contains
details about the fitted (hierarchical) Hidden Markov model.
}
