% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{fHMM_sdds}
\alias{fHMM_sdds}
\title{Define state-dependent distributions.}
\usage{
fHMM_sdds(sdds)
}
\arguments{
\item{sdds}{A character or a character vector of length two that can be specified for
\code{"sdds"} in \code{link{set_controls}}.}
}
\value{
A list of length \code{length(sdds)}. Each element is a list, containing
\itemize{
\item the \code{"name"} of the distribution
\item and a list \code{"pars"} of its parameters.
}
Unknown parameters are set to \code{NA}.
}
\description{
This function defines state-dependent distributions of an (H)HMM.
}
\examples{
fHMM_sdds(sdds = "t")
fHMM_sdds(sdds = "gamma")
fHMM_sdds(sdds = c("t(mu = 1, sigma = 2, df = 3)","gamma(sigma = 1)"))
}
