% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_sdds.R
\name{fHMM_sdds}
\alias{fHMM_sdds}
\alias{print.fHMM_sdds}
\title{Define state-dependent distributions}
\usage{
fHMM_sdds(sdds, states)

\method{print}{fHMM_sdds}(x, ...)
}
\arguments{
\item{sdds}{A \code{character}, specifying the state-dependent distribution. One of
\itemize{
\item \code{"normal"} (the normal distribution),
\item \code{"lognormal"} (the log-normal distribution),
\item \code{"t"} (the t-distribution),
\item \code{"gamma"} (the gamma distribution),
\item \code{"poisson"} (the Poisson distribution).
}

The distribution parameters, i.e. the
\itemize{
\item mean \code{mu},
\item standard deviation \code{sigma} (not for the Poisson distribution),
\item degrees of freedom \code{df} (only for the t-distribution),
}
can be fixed via, e.g., \code{"t(df = 1)"} or
\code{"gamma(mu = 0, sigma = 1)"}.
To fix different values of a parameter for different states, separate by
"|", e.g. \code{"poisson(mu = 1|2|3)"}.

If \code{hierarchy = TRUE}, \code{sdds} must be a \code{vector} of length 2.
The first entry corresponds to the coarse-scale layer, while the second entry
corresponds to the fine-scale layer.

By default, \code{sdds = "normal"} if \code{hierarchy = FALSE} and
\code{sdds = c("normal", "normal")} if \code{hierarchy = TRUE}.

See also \code{\link{fHMM_sdds}} for more details.}

\item{states}{An \code{integer}, the number of states of the underlying Markov chain.

If \code{hierarchy = TRUE}, \code{states} must be a \code{vector} of length
2. The first entry corresponds to the coarse-scale layer, while the second
entry corresponds to the fine-scale layer.

By default, \code{states = 2} if \code{hierarchy = FALSE} and
\code{states = c(2, 2)} if \code{hierarchy = TRUE}.}

\item{...}{Currently not used.}

\item{true_parameters}{An object of class \code{fHMM_parameters}, used as simulation parameters.
By default, \code{true_parameters = NULL}, i.e., sampled true parameters.}
}
\value{
An object of class \code{"fHMM_sdds"}, which is a \code{list} of length
\code{length(sdds)}, i.e., of length 2 in the hierarchical case and 1 else.

Each \code{list} element again is a \code{list}, containing the output of
\code{\link{decode_sdd}}.
}
\description{
This helper function defines state-dependent distributions for the \{fHMM\}
package.
}
\examples{
\dontrun{
fHMM_sdds(
  sdds = c("t(sigma = 0.1 | 1, df = Inf)", "gamma"),
  states = c(2, 3)
)
}

}
\keyword{internal}
