% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_closest_year}
\alias{find_closest_year}
\title{Find closest year}
\usage{
find_closest_year(date)
}
\arguments{
\item{date}{[\code{character(1)}]\cr
The date in format \code{"YYYY-MM-DD"}.}
}
\value{
An \code{integer}, the closest year to the input date.
}
\description{
This function takes a \code{date} as input and returns the closest year.
}
\keyword{internal}
