% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Model fitting}
\usage{
fit_model(data, ncluster = 1, seed = NULL)
}
\arguments{
\item{data}{An object of class \code{fHMM_data}.}

\item{ncluster}{Set the number of clusters for parallelization.}

\item{seed}{Set a seed for the sampling of initial values.}
}
\value{
An object of class \code{fHMM_model}.
}
\description{
This function fits an (hierarchical) HMM for the fHMM package.
}
