% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Model fitting for the fHMM package.}
\usage{
fit_model(data, seed = NULL)
}
\arguments{
\item{data}{An object of class \code{fHMM_data}.}

\item{seed}{Set a seed for the sampling of initial values.}
}
\value{
An object of class \code{fHMM_model}, which is a list of
\itemize{
\item ...
}
}
\description{
This function fits an (hierarchical) HMM for the fHMM package.
}
