% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{Gamma2delta}
\alias{Gamma2delta}
\title{This function computes the stationary distribution of a transition
probability matrix \code{Gamma}.}
\usage{
Gamma2delta(Gamma)
}
\arguments{
\item{Gamma}{A transition probability \code{matrix}.}
}
\value{
A stationary distribution \code{vector}.
}
\description{
This function computes the stationary distribution of a transition
probability matrix \code{Gamma}.
}
\details{
If the stationary distribution vector cannot be computed, it is set to the
discrete uniform distribution over the states.
}
\keyword{internal}
