% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{Gamma2gammasCon}
\alias{Gamma2gammasCon}
\title{This function constrains the non-diagonal matrix elements of a transition
probability matrix \code{Gamma}.}
\usage{
Gamma2gammasCon(Gamma, shift = 0.001)
}
\arguments{
\item{Gamma}{A transition probability matrix.}

\item{shift}{A numeric value for shifting boundary probabilities.}
}
\value{
A vector of constrained non-diagonal matrix elements (column-wise).
}
\description{
This function constrains the non-diagonal matrix elements of a transition
probability matrix \code{Gamma}.
}
\keyword{internal}
