% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma2gammasCon.R
\name{Gamma2gammasCon}
\alias{Gamma2gammasCon}
\title{Get constrained non-diagonal matrix elements of transition probability matrix}
\usage{
Gamma2gammasCon(Gamma)
}
\arguments{
\item{Gamma}{Transition probability matrix}
}
\value{
Constrained non-diagonal matrix elements (column-wise)
}
\description{
Get constrained non-diagonal matrix elements of transition probability matrix
}
