% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma2gammasUncon.R
\name{Gamma2gammasUncon}
\alias{Gamma2gammasUncon}
\title{Unconstraines non-diagonal matrix elements of transition probability matrix.}
\usage{
Gamma2gammasUncon(Gamma)
}
\arguments{
\item{Gamma}{Transition probability matrix.}
}
\value{
Vector of unconstrained non-diagonal matrix elements (column-wise).
}
\description{
Unconstraines non-diagonal matrix elements of transition probability matrix.
}
