% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{Gamma2gammasUncon}
\alias{Gamma2gammasUncon}
\title{This function un-constrains the non-diagonal matrix elements of a transition
probability matrix \code{Gamma}.}
\usage{
Gamma2gammasUncon(Gamma)
}
\arguments{
\item{Gamma}{A transition probability matrix.}
}
\value{
A vector of un-constrained non-diagonal matrix elements (column-wise).
}
\description{
This function un-constrains the non-diagonal matrix elements of a transition
probability matrix \code{Gamma}.
}
\keyword{internal}
