% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasCon2Gamma.R
\name{gammasCon2Gamma}
\alias{gammasCon2Gamma}
\title{Builds transition probability matrix from constrained non-diagonal elements.}
\usage{
gammasCon2Gamma(gammasCon, dim)
}
\arguments{
\item{gammasCon}{Vector of constrained non-diagonal elements of transition probability matrix.}

\item{dim}{Numeric, dimension of transition probability matrix.}
}
\value{
Transition probability matrix.
}
\description{
Builds transition probability matrix from constrained non-diagonal elements.
}
