% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{gammasCon2Gamma}
\alias{gammasCon2Gamma}
\title{This function builds a transition probability matrix of dimension \code{dim}
from constrained non-diagonal elements \code{gammasCon}.}
\usage{
gammasCon2Gamma(gammasCon, dim)
}
\arguments{
\item{gammasCon}{A vector of constrained non-diagonal elements of a transition probability
matrix.}

\item{dim}{The dimension of the transition probability matrix.}
}
\value{
A transition probability matrix.
}
\description{
This function builds a transition probability matrix of dimension \code{dim}
from constrained non-diagonal elements \code{gammasCon}.
}
\keyword{internal}
