% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasCon2gammasUncon.R
\name{gammasCon2gammasUncon}
\alias{gammasCon2gammasUncon}
\title{Unconstrains non-diagonal elements of transition probability matrix.}
\usage{
gammasCon2gammasUncon(gammasCon, dim)
}
\arguments{
\item{gammasCon}{Vector of constrained non-diagonal elements of transition probability matrix.}

\item{dim}{Numeric, dimension of transition probability matrix.}
}
\value{
Vector of unconstrained non-diagonal elements of transition probability matrix.
}
\description{
Unconstrains non-diagonal elements of transition probability matrix.
}
