% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{gammasCon2gammasUncon}
\alias{gammasCon2gammasUncon}
\title{This function un-constrains the constrained non-diagonal elements
\code{gammasCon} of a transition probability matrix of dimension \code{dim}.}
\usage{
gammasCon2gammasUncon(gammasCon, dim)
}
\arguments{
\item{gammasCon}{A vector of constrained non-diagonal elements of a transition probability
matrix.}

\item{dim}{The dimension of the transition probability matrix.}
}
\value{
A vector of un-constrained non-diagonal elements of the transition
probability matrix.
}
\description{
This function un-constrains the constrained non-diagonal elements
\code{gammasCon} of a transition probability matrix of dimension \code{dim}.
}
\keyword{internal}
