% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{gammasUncon2Gamma}
\alias{gammasUncon2Gamma}
\title{Builds transition probability matrix from unconstrained non-diagonal elements.}
\usage{
gammasUncon2Gamma(gammasUncon, dim)
}
\arguments{
\item{gammasUncon}{Vector of unconstrained non-diagonal elements of transition probability matrix.}

\item{dim}{Numeric, dimension of transition probability matrix.}
}
\value{
Transition probability matrix.
}
\description{
Builds transition probability matrix from unconstrained non-diagonal elements.
}
\keyword{internal}
