% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{gammasUncon2gammasCon}
\alias{gammasUncon2gammasCon}
\title{This function constrains non-diagonal elements \code{gammasUncon} of a
transition probability matrix.}
\usage{
gammasUncon2gammasCon(gammasUncon, dim)
}
\arguments{
\item{gammasUncon}{A vector of un-constrained non-diagonal elements of a transition probability
matrix.}

\item{dim}{The dimension of the transition probability matrix.}
}
\value{
A vector of constrained non-diagonal elements of a transition probability
matrix.
}
\description{
This function constrains non-diagonal elements \code{gammasUncon} of a
transition probability matrix.
}
\keyword{internal}
