% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammasUncon2gammasCon.R
\name{gammasUncon2gammasCon}
\alias{gammasUncon2gammasCon}
\title{Constrains non-diagonal elements of transition probability matrix.}
\usage{
gammasUncon2gammasCon(gammasUncon, dim)
}
\arguments{
\item{gammasUncon}{Vector of unconstrained non-diagonal elements of transition probability matrix.}

\item{dim}{Numeric, dimension of transition probability matrix.}
}
\value{
Vector of constrained non-diagonal elements of transition probability matrix.
}
\description{
Constrains non-diagonal elements of transition probability matrix.
}
