% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tpm}
\alias{is_tpm}
\title{Check for transition probability matrix}
\usage{
is_tpm(x)
}
\arguments{
\item{x}{A \code{matrix}.}
}
\value{
A \code{logical}.
}
\description{
This function checks if \code{x} is a tpm (transition probability matrix).
}
\examples{
\dontrun{
is_tpm(diag(2))
is_tpm(matrix(1, 2, 2))
}
}
\keyword{internal}
\keyword{utils}
