% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_tpm}
\alias{is_tpm}
\title{Check for tpm}
\usage{
is_tpm(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A boolean.
}
\description{
This function checks if \code{x} is a tpm (transition probability matrix).
}
\examples{
fHMM:::is_tpm(diag(2))
fHMM:::is_tpm(matrix(1, 2, 2))
}
\keyword{utils}
