% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{list_to_vector}
\alias{list_to_vector}
\title{List to vector}
\usage{
list_to_vector(x)
}
\arguments{
\item{x}{[\code{list()}]\cr
A \code{list}.}
}
\value{
A \code{numeric}.
}
\description{
This function produces a \code{vector} from a \code{list} structure and
replaces \code{NULL} elements by \code{NA}.
}
\keyword{internal}
