% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_all}
\alias{match_all}
\title{Best-possible match of two numeric vectors}
\usage{
match_all(x, y)
}
\arguments{
\item{x}{A \code{numeric} vector.}

\item{y}{Another \code{numeric} vector of the same length as \code{x}.}
}
\value{
An \code{integer} vector of length \code{length(x)} with the positions of \code{y}
in \code{x}.
}
\description{
This function matches the positions of two numeric vectors as good as possible.
}
\examples{
\dontrun{
x <- c(-1, 0, 1)
y <- c(0.1, 2, -1.2)
match_all(x = x, y = y)
}
}
\keyword{internal}
\keyword{utils}
