% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_likelihood.R
\name{max_likelihood}
\alias{max_likelihood}
\title{Maximize the model's log-likelihood function}
\usage{
max_likelihood(data, controls)
}
\arguments{
\item{data}{A list of processed data information}

\item{controls}{A list of controls}
}
\value{
A fitted model
}
\description{
Maximize the model's log-likelihood function
}
