% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{muCon2muUncon}
\alias{muCon2muUncon}
\title{Unconstrains expected values.}
\usage{
muCon2muUncon(muCon, link)
}
\arguments{
\item{muCon}{Vector of constrained expected values.}

\item{link}{Boolean, determining whether to apply the link function.}
}
\value{
Vector of unconstrained expected values.
}
\description{
Unconstrains expected values.
}
\keyword{internal}
