% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{muCon2muUncon}
\alias{muCon2muUncon}
\title{This function un-constrains the constrained expected values \code{muCon}.}
\usage{
muCon2muUncon(muCon, link)
}
\arguments{
\item{muCon}{A vector of constrained expected values.}

\item{link}{A boolean, determining whether to apply the link function.}
}
\value{
A vector of un-constrained expected values.
}
\description{
This function un-constrains the constrained expected values \code{muCon}.
}
\keyword{internal}
