% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muCon2muUncon.R
\name{muCon2muUncon}
\alias{muCon2muUncon}
\title{Unconstrain expected values}
\usage{
muCon2muUncon(muCon, link)
}
\arguments{
\item{muCon}{Constrained expected values}

\item{link}{Boolean determining whether to apply the link function}
}
\value{
Unconstrained expected values
}
\description{
Unconstrain expected values
}
