% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{muUncon2muCon}
\alias{muUncon2muCon}
\title{This function constrains the un-constrained expected values \code{muUncon}.}
\usage{
muUncon2muCon(muUncon, link)
}
\arguments{
\item{muUncon}{A vector of un-constrained expected values.}

\item{link}{A boolean, determining whether to apply the link function.}
}
\value{
A vector of constrained expected values.
}
\description{
This function constrains the un-constrained expected values \code{muUncon}.
}
\keyword{internal}
