% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muUncon2muCon.R
\name{muUncon2muCon}
\alias{muUncon2muCon}
\title{Constrain expected values}
\usage{
muUncon2muCon(muUncon, link)
}
\arguments{
\item{muUncon}{Unconstrained expected values}

\item{link}{Boolean determining whether to apply the link function}
}
\value{
Constrained expected values
}
\description{
Constrain expected values
}
