% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_likelihood.R
\name{nLL_hhmm}
\alias{nLL_hhmm}
\title{Negative log-likelihood of a HHMM}
\usage{
nLL_hhmm(thetaUncon, observations, controls)
}
\arguments{
\item{thetaUncon}{Vector of model parameters in format \code{thetaUncon}.}

\item{observations}{A matrix of observations, coarse-scale observations in the first column and fine-scale observations in the corresponding row.}

\item{controls}{A list of controls.}
}
\value{
Numeric, negative log-likelihood value.
}
\description{
Computes the negative log-likelihood of a hierarchical hidden Markov model.
}
\keyword{internal}
