% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLL_hhmm.R
\name{nLL_hhmm}
\alias{nLL_hhmm}
\title{Compute the negative log-likelihood of a hierarchical hidden Markov model}
\usage{
nLL_hhmm(thetaUncon, observations, controls)
}
\arguments{
\item{thetaUncon}{Model parameters in format \code{thetaUncon}}

\item{observations}{A vector of observations}

\item{controls}{A list of controls}
}
\value{
Negative log-likelihood
}
\description{
Compute the negative log-likelihood of a hierarchical hidden Markov model
}
