% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_likelihood.R
\name{nLL_hhmm}
\alias{nLL_hhmm}
\title{Negative log-likelihood function of an HHMM}
\usage{
nLL_hhmm(parUncon, observations, controls)
}
\arguments{
\item{parUncon}{An object of class \code{parUncon}.}

\item{observations}{The matrix of the simulated or empirical data used for estimation.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
The negative log-likelihood value.
}
\description{
This function computes the negative log-likelihood of an HHMM.
}
\keyword{internal}
