% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nLL_hmm.R
\name{nLL_hmm}
\alias{nLL_hmm}
\title{Negative log-likelihood of a HMM}
\usage{
nLL_hmm(thetaUncon, observations, controls)
}
\arguments{
\item{thetaUncon}{Vector of model parameters in format \code{thetaUncon}.}

\item{observations}{A vector of observations.}

\item{controls}{A list of controls.}
}
\value{
Numeric, negative log-likelihood value.
}
\description{
Computes the negative log-likelihood of a hidden Markov model.
}
