% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{nLL_hmm}
\alias{nLL_hmm}
\title{Negative log-likelihood function of an HMM}
\usage{
nLL_hmm(parUncon, observations, controls)
}
\arguments{
\item{parUncon}{An object of class \code{parUncon}.}

\item{observations}{The vector of the simulated or empirical data used for estimation.}

\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
The negative log-likelihood value.
}
\description{
This function computes the negative log-likelihood of an HMM.
}
\keyword{internal}
