% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_labels.R
\name{parameter_labels}
\alias{parameter_labels}
\title{Create labels for estimated parameters}
\usage{
parameter_labels(controls, expected_length)
}
\arguments{
\item{controls}{An object of class \code{fHMM_controls}.}

\item{expected_length}{The expected output length.}
}
\value{
A character vector of parameter labels.
}
\description{
This function creates labels for the estimated HMM parameters.
}
\examples{
fHMM:::parameter_labels(set_controls(), 8)
}
\keyword{internal}
