% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{parameter_transformations}
\alias{parameter_transformations}
\alias{par2parUncon}
\alias{parUncon2parCon}
\alias{parCon2par}
\alias{par2parCon}
\alias{parCon2parUncon}
\alias{parUncon2par}
\alias{muCon2muUncon}
\alias{muUncon2muCon}
\alias{sigmaCon2sigmaUncon}
\alias{sigmaUncon2sigmaCon}
\alias{dfCon2dfUncon}
\alias{dfUncon2dfCon}
\alias{Gamma2gammasCon}
\alias{Gamma2gammasUncon}
\alias{gammasCon2Gamma}
\alias{gammasCon2gammasUncon}
\alias{gammasUncon2Gamma}
\alias{gammasUncon2gammasCon}
\title{Parameter transformations}
\usage{
par2parUncon(par, controls)

parUncon2parCon(parUncon, controls)

parCon2par(parCon, controls)

par2parCon(par, controls)

parCon2parUncon(parCon, controls)

parUncon2par(parUncon, controls)

muCon2muUncon(muCon, link)

muUncon2muCon(muUncon, link)

sigmaCon2sigmaUncon(sigmaCon)

sigmaUncon2sigmaCon(sigmaUncon)

dfCon2dfUncon(dfCon)

dfUncon2dfCon(dfUncon)

Gamma2gammasCon(Gamma, shift = 0.001)

Gamma2gammasUncon(Gamma)

gammasCon2Gamma(gammasCon, dim)

gammasCon2gammasUncon(gammasCon, dim)

gammasUncon2Gamma(gammasUncon, dim)

gammasUncon2gammasCon(gammasUncon, dim)
}
\arguments{
\item{par}{An object of class \code{\link{fHMM_parameters}}.}

\item{controls}{Either a \code{list} or an object of class \code{fHMM_controls}.

The \code{list} can contain the following elements, which are described
in more detail below:
\itemize{
\item \code{hierarchy}, defines an hierarchical HMM,
\item \code{states}, defines the number of states,
\item \code{sdds}, defines the state-dependent distributions,
\item \code{horizon}, defines the time horizon,
\item \code{period}, defines a flexible, periodic fine-scale time horizon,
\item \code{data}, a \code{list} of controls that define the data,
\item \code{fit}, a \code{list} of controls that define the model fitting,
\item \code{seed}, defines a seed for reproducibility,
\item \code{verbose}, defines whether to print progress messages.
}
Either none, all, or selected elements can be specified.

Unspecified parameters are set to their default values, see below.

Specifications in \code{controls} override individual specifications.}

\item{muCon, muUncon}{A vector of (un-) constrained expected values.}

\item{link}{Either \code{TRUE} or \code{FALSE}, determining whether to apply the link
function.}

\item{sigmaCon, sigmaUncon}{A vector of (un-) constrained standard deviations.}

\item{dfCon, dfUncon}{A vector of (un-) constrained degrees of freedom.}

\item{shift}{A small, positive \code{numeric} for shifting boundary probabilities.
By default, \code{shift = 1e-3}.}

\item{gammasCon, gammasUncon}{A vector of (un-) constrained non-diagonal transition probabilities.}

\item{dim}{An \code{integer}, the dimension of the transition probability matrix.}
}
\value{
For \code{par2parUncon}: a vector of unconstrained model parameters.

For \code{parUncon2parCon}: a vector of constrained model parameters.

For \code{parCon2par}: an object of class \code{\link{fHMM_parameters}}.

For \code{par2parCon}: a vector of constrained model parameters.

For \code{parCon2parUncon}: a vector of unconstrained model parameters.

For \code{parUncon2par}: an object of class \code{fHMM_parameters}.

For \code{muCon2muUncon}: a vector of unconstrained expected values.

For \code{muUncon2muCon}: a vector of constrained expected values.

For \code{sigmaCon2sigmaUncon}: a vector of unconstrained standard
deviations.

For \code{sigmaUncon2sigmaCon}: a vector of constrained standard deviations.

For \code{dfCon2dfUncon}: a vector of unconstrained degrees of freedom.

For \code{dfUncon2dfCon}: a vector of constrained degrees of freedom.

For \code{Gamma2gammasCon}: a vector of constrained non-diagonal matrix
elements (column-wise).

For \code{Gamma2gammasUncon}: a vector of unconstrained non-diagonal matrix
elements (column-wise).

For \code{gammasCon2Gamma}: a transition probability matrix.

For \code{gammasCon2gammasUncon}: a vector of unconstrained non-diagonal
elements of the transition probability matrix.

For \code{gammasUncon2Gamma}: a transition probability matrix.

For \code{gammasUncon2gammasCon}: a vector of constrained non-diagonal
elements of a transition probability matrix.
}
\description{
These helper functions transform model parameters between constrained
(suffix \code{*Con}) and unconstrained spaces
(suffix \code{*Uncon}) for numerical optimization.
}
\keyword{internal}
