% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fHMM_data}
\alias{plot.fHMM_data}
\title{Plot method for an object of class \code{fHMM_data}}
\usage{
\method{plot}{fHMM_data}(x, events = NULL, title = NULL, from = NULL, to = NULL, ...)
}
\arguments{
\item{x}{[\code{fHMM_data}]\cr
An object of class \code{fHMM_data}.}

\item{events}{[\code{NULL} | \code{fHMM_events}]\cr
An object of class \code{\link{fHMM_events}}.}

\item{title}{[\code{NULL} | \code{character(1)}]\cr
Optionally a \code{character} for a custom title.}

\item{from}{[\code{NULL} | \code{character(1)}]\cr
Optionally a \code{character}, a date in format \code{"YYYY-MM-DD"},
setting the lower date bound for plotting.
By default, \code{from = NULL}, i.e. no lower bound.}

\item{to}{[\code{NULL} | \code{character(1)}]\cr
Optionally a \code{character}, a date in format \code{"YYYY-MM-DD"},
setting the upper date bound for plotting.
By default, \code{to = NULL}, i.e. no upper bound.}

\item{...}{Currently not used.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{fHMM_data}.
}
\examples{
plot(dax_model_3t$data, title = "DAX time series")

}
