% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fHMM_data.R
\name{plot.fHMM_data}
\alias{plot.fHMM_data}
\title{Plot method for \code{fHMM_data}.}
\usage{
\method{plot}{fHMM_data}(x, events = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_data}.}

\item{events}{...}

\item{...}{Ignored.}
}
\value{
No return value.
}
\description{
This function is the plot method for an object of class \code{fHMM_data}.
}
