% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fHMM_data}
\alias{plot.fHMM_data}
\title{Plot method for an object of class \code{fHMM_data}}
\usage{
\method{plot}{fHMM_data}(x, events = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_data}.}

\item{events}{Either \code{NULL} or an object of class \code{fHMM_events}.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{fHMM_data}.
}
