% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fHMM_model}
\alias{plot.fHMM_model}
\title{Plot method for an object of class \code{fHMM_model}}
\usage{
\method{plot}{fHMM_model}(x, plot_type = "ts", events = NULL, colors = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{fHMM_model}.}

\item{plot_type}{A character (vector), specifying the type of plot and can be one (or more) of
\itemize{
\item \code{"ll"} for a visualization of the likelihood values in the
different optimization runs,
\item \code{"sdds"} for a visualization of the estimated state-dependent
distributions,
\item \code{"pr"} for a visualization of the model's (pseudo-) residuals,
\item \code{"ts"} for a visualization of the financial time series.
}}

\item{events}{An object of class \code{fHMM_events}.}

\item{colors}{Either \code{NULL} or a character vector of color names or hexadecimal RGB
triplets.}

\item{...}{Ignored.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function is the plot method for an object of class \code{fHMM_model}.
}
