% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ll}
\alias{plot_ll}
\title{Visualization of log-likelihood values}
\usage{
plot_ll(lls, ll_relative = TRUE)
}
\arguments{
\item{lls}{[\code{numeric()}]\cr
A \code{numeric} vector of log-likelihood values.}

\item{ll_relative}{[\code{logical(1)}]\cr
A \code{logical}, set to \code{TRUE} (default) to plot the differences from
the best log-likelihood value. Set to \code{FALSE} to plot the absolute
values.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots the log-likelihood values of the different optimization
runs.
}
\keyword{internal}
