% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ll.R
\name{plot_ll}
\alias{plot_ll}
\title{Visualization of log-likelihood values}
\usage{
plot_ll(lls, controls)
}
\arguments{
\item{lls}{A vector of log-likelihood values.}

\item{controls}{A list of controls.}
}
\value{
No return value. Creates file "log_likelihoods.pdf" in "\code{controls[["path"]]}/models/\code{controls[["id"]]}".
}
\description{
Plots log-likelihood values of the different estimation runs.
}
