% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ll}
\alias{plot_ll}
\title{Visualization of log-likelihood values}
\usage{
plot_ll(lls)
}
\arguments{
\item{lls}{A numeric vector of log-likelihood values.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function plots the log-likelihood values of the different estimation runs.
}
\keyword{internal}
