% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ll.R
\name{plot_ll}
\alias{plot_ll}
\title{Plot log-likelihood values of estimation runs}
\usage{
plot_ll(llks, controls)
}
\arguments{
\item{llks}{A vector of log-likelihood values}

\item{controls}{A list of controls}
}
\value{
No return value, creates graphic in \code{controls[["path"]]}/models/\code{controls[["id"]]}
}
\description{
Plot log-likelihood values of estimation runs
}
