% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_pr}
\alias{plot_pr}
\title{Visualize pseudo residuals}
\usage{
plot_pr(residuals, hierarchy)
}
\arguments{
\item{residuals}{[\code{fHMM_residuals}]\cr
An object of class \code{fHMM_residuals}.}

\item{hierarchy}{[\code{logical(1)}]\cr
The element \code{controls$hierarchy}.}
}
\value{
No return value. Draws a plot to the current device.
}
\description{
This function visualizes the pseudo residuals.
}
\keyword{internal}
