% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdd.R
\name{plot_sdd}
\alias{plot_sdd}
\title{Visualize estimated state-dependent distributions}
\usage{
plot_sdd(controls, data, fit, decoding, colors)
}
\arguments{
\item{controls}{A list of controls}

\item{data}{A list of processed data information}

\item{fit}{A fitted model}

\item{decoding}{A matrix of decoded states}

\item{colors}{A matrix of colors for different states}
}
\value{
No return value, creates graphic in \code{controls[["path"]]}/models/\code{controls[["id"]]}
}
\description{
Visualize estimated state-dependent distributions
}
