% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sdds.R
\name{plot_sdds}
\alias{plot_sdds}
\title{Visualization of estimated state-dependent distributions.}
\usage{
plot_sdds(estimated_parameters, true_parameters, colors)
}
\arguments{
\item{estimated_parameters}{An object of class \code{fHMM_parameters}.}

\item{true_parameters}{Either \code{NULL} or an object of class \code{fHMM_parameters}.}

\item{colors}{An object of class \code{fHMM_colors}.}
}
\value{

}
\description{
This function plots the estimated state-dependent distributions.
}
