% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Visualize data time series.}
\usage{
plot_ts(data, decoding, colors, events, predict)
}
\arguments{
\item{data}{A list of processed data information.}

\item{decoding}{A matrix of decoded states.}

\item{colors}{An object of class \code{fHMM_colors}. Ignored if \code{decoding = NULL}.}

\item{events}{A list of events.}

\item{predict}{An object of class \code{fHMM_predict}.}

\item{This}{function visualizes the data time series. If \code{decoding} is
specified, the decoding is highlighted. If \code{events} is specified,
the events are highlighted.}
}
\value{
...
}
\description{
Visualize data time series.
}
