% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Visualize decoded time-series}
\usage{
plot_ts(controls, data, decoding, colors, events)
}
\arguments{
\item{controls}{A list of controls.}

\item{data}{A list of processed data information.}

\item{decoding}{A matrix of decoded states.}

\item{colors}{A matrix of colors for different states.}

\item{events}{A list of events.}
}
\value{
No return value, creates graphic in \code{controls[["path"]]}/models/\code{controls[["id"]]}
}
\description{
Visualize decoded time-series
}
