% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fHMM_model.R
\name{predict.fHMM_model}
\alias{predict.fHMM_model}
\title{Prediction}
\usage{
\method{predict}{fHMM_model}(object, ahead, ci_level = 0.05, ...)
}
\arguments{
\item{object}{An object of class \code{fHMM_model}.}

\item{ahead}{A positive integer.}

\item{ci_level}{The confidence level, a numeric between 0 and 1.}

\item{...}{Ignored.}
}
\value{
An object of class \code{fHMM_model}.
}
\description{
This function...
}
