% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(controls, true_parameters = NULL, seed = NULL)
}
\arguments{
\item{controls}{[\code{fHMM_controls}]\cr
An object of class \code{fHMM_controls}.}

\item{true_parameters}{[\code{NULL} | \code{fHMM_parameters}]\cr
An object of class \code{fHMM_parameters}, used as simulation parameters.
By default, \code{true_parameters = NULL}, i.e., sampled true parameters.}

\item{seed}{[\code{NULL} | \code{integer(1)}]\cr
Set a seed for the data simulation.
No seed by default.}
}
\value{
An object of class \code{\link{fHMM_data}}.
}
\description{
This function simulates or reads financial data for the \{fHMM\} package.
}
\examples{
controls <- set_controls()
data <- prepare_data(controls)
class(data)
summary(data)

}
