% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.fHMM_model.R
\name{print.summary.fHMM_model}
\alias{print.summary.fHMM_model}
\title{Print method for the summary of \code{fHMM_model}.}
\usage{
\method{print}{summary.fHMM_model}(x)
}
\arguments{
\item{x}{An object of class \code{summary.fHMM_model}.}

\item{...}{Ignored.}
}
\value{
...
}
\description{
Print method for the summary of \code{fHMM_model}.
}
