% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Process data}
\usage{
process_data(controls, sim_par = NULL)
}
\arguments{
\item{controls}{A list of controls}

\item{sim_par}{A vector of model parameters for simulation, default \code{NULL}}
}
\value{
A list of processed data information
}
\description{
Process data
}
