% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_residuals.R
\name{pseudo_residuals}
\alias{pseudo_residuals}
\title{Compute and visualize pseudo-residuals}
\usage{
pseudo_residuals(controls, data, fit, decoding)
}
\arguments{
\item{controls}{A list of controls}

\item{data}{A list of processed data information}

\item{fit}{A fitted model}

\item{decoding}{A matrix of decoded states}
}
\value{
No return value, creates graphic in \code{controls[["path"]]}/models/\code{controls[["id"]]}
}
\description{
Compute and visualize pseudo-residuals
}
