% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read financial data from .csv-file}
\usage{
read_data(controls)
}
\arguments{
\item{controls}{A list of controls}
}
\value{
A list containing the following elements:
\item{logReturns}{Log-returns, the data that is modeled}
\item{dataRaw}{Raw data}
\item{dataRaw_cs}{Raw data on coarse-scale}
\item{dates}{Vector of dates}
\item{T_star}{Vector of fine-scale chunk sizes}
}
\description{
Read financial data from .csv-file
}
