% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read .csv-file}
\usage{
read_data(controls)
}
\arguments{
\item{controls}{A list of controls.}
}
\value{
A list containing the following elements:
\item{logReturns}{log-returns (the data that is modeled)}
\item{dataRaw}{raw data}
\item{dataRaw_cs}{raw coarse-scale data}
\item{dates}{vector of dates}
\item{T_star}{vector of fine-scale chunk sizes}
}
\description{
Reads financial data from .csv-file.
}
