% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data}
\usage{
read_data(controls)
}
\arguments{
\item{controls}{An object of class \code{fHMM_controls}.}
}
\value{
A \code{list} containing the following elements:
\itemize{
\item the \code{matrix} of the \code{dates} if \code{controls$simulated = FALSE}
and \code{controls$data$data_column} is specified,
\item the \code{matrix} of the \code{time_points} if \code{controls$simulated = TRUE}
or \code{controls$data$data_column} is not specified,
\item the \code{matrix} of the empirical \code{data} used for estimation,
\item the \code{matrix} named \code{time_series} of empirical data before
the transformation to log-returns,
\item the \code{vector} of fine-scale chunk sizes \code{T_star} if
\code{controls$hierarchy = TRUE}.
}
}
\description{
This helper function reads financial data for the \{fHMM\} package.
}
\keyword{internal}
