% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read .csv-file}
\usage{
read_data(controls)
}
\arguments{
\item{controls}{A list of controls.}
}
\value{
A list containing the following elements:
\item{data}{A matrix of data that is modeled.}
\item{data_raw}{A matrix of raw data.}
\item{data_fs_raw}{A matrix of raw fine-scale data.}
\item{data_cs_raw}{A matrix of raw coarse-scale data.}
\item{dates}{A vector of dates.}
\item{T_star}{A vector of fine-scale chunk sizes.}
}
\description{
Reads financial data from .csv-file.
}
