% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_tpm}
\alias{sample_tpm}
\title{Sample transition probability matrices}
\usage{
sample_tpm(dim)
}
\arguments{
\item{dim}{The matrix dimension.}
}
\value{
A transition probability matrix.
}
\description{
This function returns a random, squared matrix of dimension \code{dim}
that fulfills the properties of a transition probability matrix.
}
\examples{
fHMM:::sample_tpm(dim = 3)
}
\keyword{utils}
