% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fHMM_parameters.R
\name{sigmaCon2sigmaUncon}
\alias{sigmaCon2sigmaUncon}
\title{This function un-constrains the constrained standard deviations
\code{sigmaCon}.}
\usage{
sigmaCon2sigmaUncon(sigmaCon)
}
\arguments{
\item{sigmaCon}{A vector of constrained standard deviations.}
}
\value{
A vector of un-constrained standard deviations.
}
\description{
This function un-constrains the constrained standard deviations
\code{sigmaCon}.
}
\keyword{internal}
