% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaCon2sigmaUncon.R
\name{sigmaCon2sigmaUncon}
\alias{sigmaCon2sigmaUncon}
\title{Unconstrains standard deviations.}
\usage{
sigmaCon2sigmaUncon(sigmaCon)
}
\arguments{
\item{sigmaCon}{Vector of constrained standard deviations.}
}
\value{
Vector of unconstrained standard deviations.
}
\description{
Unconstrains standard deviations.
}
