% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaUncon2sigmaCon.R
\name{sigmaUncon2sigmaCon}
\alias{sigmaUncon2sigmaCon}
\title{Constrain standard deviations}
\usage{
sigmaUncon2sigmaCon(sigmaUncon)
}
\arguments{
\item{sigmaUncon}{Unconstrained standard deviations}
}
\value{
Constrained standard deviations
}
\description{
Constrain standard deviations
}
