% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaUncon2sigmaCon.R
\name{sigmaUncon2sigmaCon}
\alias{sigmaUncon2sigmaCon}
\title{Constrains standard deviations.}
\usage{
sigmaUncon2sigmaCon(sigmaUncon)
}
\arguments{
\item{sigmaUncon}{Vector of unconstrained standard deviations.}
}
\value{
Vector of constrained standard deviations.
}
\description{
Constrains standard deviations.
}
