% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data}
\usage{
simulate_data(controls, true_parameters, seed = NULL)
}
\arguments{
\item{controls}{An object of class \code{fHMM_controls}.}

\item{true_parameters}{An object of class \code{fHMM_parameters}, used as simulation parameters.}

\item{seed}{Set a seed for the data simulation.}
}
\value{
A list containing the following elements:
\itemize{
\item the matrix of \code{time_points},
\item the matrix of the simulated \code{markov_chain},
\item the matrix of the simulated \code{data},
\item the vector of fine-scale chunk sizes \code{T_star} if \code{controls$hierarchy = TRUE}.
}
}
\description{
This function simulates data for the {fHMM} package.
}
\keyword{internal}
