% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Data simulation}
\usage{
simulate_data(controls, sim_par = NA)
}
\arguments{
\item{controls}{A list of controls.}

\item{sim_par}{A list of model parameters for simulation in \code{thetaList} format, default \code{NA}.}
}
\value{
A list containing the following elements:
\item{logReturns}{simulated log-returns}
\item{states0}{simulated hidden states}
\item{thetaUncon0}{true parameters in format \code{thetaUncon}}
\item{thetaCon0}{true parameters in format \code{thetaCon}}
\item{thetaList0}{true parameters in format \code{thetaList}}
\item{T_star}{vector of fine-scale chunk sizes}
}
\description{
Simulates data from a (hierarchical) hidden Markov model.
}
