% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm.R
\name{simulate_hmm}
\alias{simulate_hmm}
\title{Simulate data}
\usage{
simulate_hmm(
  controls = list(),
  hierarchy = FALSE,
  states = if (!hierarchy) 2 else c(2, 2),
  sdds = if (!hierarchy) "normal" else c("normal", "normal"),
  horizon = if (!hierarchy) 100 else c(100, 30),
  period = if (hierarchy && is.na(horizon[2])) "m" else NA,
  true_parameters = fHMM_parameters(controls = controls, hierarchy = hierarchy, states =
    states, sdds = sdds),
  seed = NULL
)
}
\arguments{
\item{controls}{An object of class \code{fHMM_controls}.}

\item{true_parameters}{An object of class \code{fHMM_parameters}, used as simulation parameters.
By default, \code{true_parameters = NULL}, i.e., sampled true parameters.}

\item{seed}{Set a seed for the data simulation.
No seed per default.}
}
\value{
A \code{list} containing the following elements:
\itemize{
\item \code{time_points}, the \code{vector} (or \code{matrix} in the
hierarchical case) of time points,
\item \code{markov_chain}, the \code{vector} (or \code{matrix} in the
hierarchical case) of the simulated states,
\item \code{data}, the \code{vector} (or \code{matrix} in the hierarchical
case) of the simulated state-dependent observations,
\item \code{T_star}, the \code{numeric} vector of fine-scale chunk sizes
in the hierarchical case
}
}
\description{
This helper function simulates HMM data.
}
