% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_hmm.R
\name{simulate_markov_chain}
\alias{simulate_markov_chain}
\title{Simulate Markov chain}
\usage{
simulate_markov_chain(
  Gamma,
  T,
  delta = stationary_distribution(Gamma),
  seed = NULL
)
}
\arguments{
\item{Gamma}{A transition probability \code{matrix}.}

\item{T}{An \code{integer}, the length of the Markov chain.}

\item{delta}{A \code{numeric} probability vector, the initial distribution.
If not specified, \code{delta} is set to the stationary distribution of
\code{Gamma}.}

\item{seed}{An \code{integer}, a seed for reproducibility (e.g., for data simulation,
parameter sampling, or random initialization of optimization runs), passed
on to \code{\link{set.seed}}.

By default, \code{seed = NULL} (i.e., no seed).}
}
\value{
A \code{numeric} vector of length \code{T} with states.
}
\description{
This helper function simulates a Markov chain.
}
\examples{
\dontrun{
Gamma <- sample_tpm(dim = 3)
T <- 10
simulate_markov_chain(Gamma = Gamma, T = T)
}
}
