% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_observations}
\alias{simulate_observations}
\title{Simulate state-dependent observations}
\usage{
simulate_observations(
  markov_chain,
  sdd,
  mus,
  sigmas,
  dfs = NULL,
  seed = NULL,
  total_length = length(markov_chain)
)
}
\arguments{
\item{markov_chain}{A numeric vector of states of a Markov chain.}

\item{sdd}{The name of the state-dependent distribution, one of \code{"t"} and
\code{"gamma"}.}

\item{mus}{A vector of expected values.}

\item{sigmas}{A vector of standard deviations.}

\item{dfs}{A vector of degrees of freedom (only relevant if \code{sdd = "t"}).}

\item{seed}{Set a seed.}

\item{total_length}{An integer, the total length of the output vector. Must be greater or equal
than \code{length(markov_chain)}.}
}
\value{
A numeric vector of length \code{total_length}, where the first
\code{length(markov_chain)} elements are numeric values and the last
\code{total_length - length(markov_chain)} elements are \code{NA}.
}
\description{
This function simulates state-dependent observations.
}
\keyword{internal}
