% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaCon2thetaList.R
\name{thetaCon2thetaList}
\alias{thetaCon2thetaList}
\title{Brings contrained model parameters from vector form \code{thetaCon} to list form \code{ThetaList}.}
\usage{
thetaCon2thetaList(thetaCon, controls)
}
\arguments{
\item{thetaCon}{Constrained model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
Constrained model parameters in list form.
}
\description{
Brings contrained model parameters from vector form \code{thetaCon} to list form \code{ThetaList}.
}
