% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thetaList2thetaListOrdered.R
\name{thetaList2thetaListOrdered}
\alias{thetaList2thetaListOrdered}
\title{Orders states in \code{thetaList} based on expected values.}
\usage{
thetaList2thetaListOrdered(thetaList, controls)
}
\arguments{
\item{thetaList}{Constrained model parameters in list form.}

\item{controls}{A list of controls.}
}
\value{
Constrained and ordered model parameters in list form.
}
\description{
Orders states in \code{thetaList} based on expected values.
}
\details{
If state-dependent distribution is \emph{t}, states are ordered decreasing with respect to expected values. If distribution is \emph{Gamma}, vice versa.
}
