% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{thetaList2thetaUncon}
\alias{thetaList2thetaUncon}
\title{Brings constrained model parameters in list form to unconstrained parameters in vector form.}
\usage{
thetaList2thetaUncon(thetaList, controls)
}
\arguments{
\item{thetaList}{Constrained model parameters in list form.}

\item{controls}{A list of controls.}
}
\value{
Unconstrained model parameters in vector form.
}
\description{
Brings constrained model parameters in list form to unconstrained parameters in vector form.
}
\keyword{internal}
