% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{thetaUncon2thetaList}
\alias{thetaUncon2thetaList}
\title{Brings uncontrained model parameters \code{thetaUncon} in list form \code{thetaList}.}
\usage{
thetaUncon2thetaList(thetaUncon, controls)
}
\arguments{
\item{thetaUncon}{Unconstrained model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
Constrained model parameters in list form.
}
\description{
Brings uncontrained model parameters \code{thetaUncon} in list form \code{thetaList}.
}
\keyword{internal}
