% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{thetaUncon2thetaUnconSplit}
\alias{thetaUncon2thetaUnconSplit}
\title{Splits uncontrained model parameters \code{thetaUncon} by fine-scale HMMs.}
\usage{
thetaUncon2thetaUnconSplit(thetaUncon, controls)
}
\arguments{
\item{thetaUncon}{Unconstrained model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
List of unconstrained fine-scale model parameters for each fine-scale HMM.
}
\description{
Splits uncontrained model parameters \code{thetaUncon} by fine-scale HMMs.
}
\keyword{internal}
