% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_transformations.R
\name{thetaUnconSplit2thetaList}
\alias{thetaUnconSplit2thetaList}
\title{Brings uncontrained fine-scale model parameters in constrained list form.}
\usage{
thetaUnconSplit2thetaList(thetaUncon, controls)
}
\arguments{
\item{thetaUncon}{Unconstrained fine-scale model parameters in vector form.}

\item{controls}{A list of controls.}
}
\value{
Constrained fine-scale model parameters in list form.
}
\description{
Brings uncontrained fine-scale model parameters in constrained list form.
}
\keyword{internal}
