% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{Unemployment rate data USA}
\format{
A \code{data.frame} with 24837 rows and the following 2 columns:
\itemize{
\item \code{date}: The date.
\item \code{rate}: The unemployment rate.
}
}
\source{
OECD (2023), Unemployment rate (indicator).
doi: 10.1787/52570002-en (Accessed on 18 January 2023)
\url{https://data.oecd.org/unemp/unemployment-rate.htm}
}
\usage{
unemp
}
\description{
The monthly unemployment rate in the USA from 1955 to 2022 on a daily
observation basis.
}
\keyword{data}
