% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{Unemployment rate data USA}
\format{
A \code{data.frame} with 24806 rows and the following 3 columns:
\itemize{
\item \code{date}: The date.
\item \code{rate}: The unemployment rate.
\item \code{rate_diff}: The difference rate to previous month.
}
}
\source{
OECD (2023), Unemployment rate (indicator).
doi: 10.1787/52570002-en (Accessed on 18 January 2023)
\url{https://www.oecd.org/en/data/indicators/unemployment-rate.html}
}
\usage{
unemp
}
\description{
The monthly unemployment rate in the USA from 1955 to 2022 on a daily
observation basis.
}
\keyword{data}
