% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_and_models.R
\docType{data}
\name{vw}
\alias{vw}
\title{Volkswagen AG (VW) stock data}
\format{
A \code{data.frame} with 6260 rows and the following 7 columns:
\itemize{
\item \code{Date}: The date.
\item \code{Open}: Opening price.
\item \code{High}: Highest price.
\item \code{Low}: Lowest price.
\item \code{Close}: Close price adjusted for splits.
\item \code{Adj.Close}: Close price adjusted for dividends and splits.
\item \code{Volume}: Trade volume.
}
}
\usage{
vw
}
\description{
Volkswagen AG (VW) stock data from 1998 to 2022 from Yahoo Finance.
}
\details{
The data was obtained via:
\preformatted{
vw <- download_data(
  symbol = "VOW3.DE",  # Volkswagen AG identifier on Yahoo Finance
  from = "1988-07-22", # first observation
  to = "2022-12-31"    # last observation
)
}
}
\keyword{data}
